% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{add_flow}
\alias{add_flow}
\title{Add one or several flows to a \code{CompartmentModel} object.}
\usage{
add_flow(model, from, to, ..., rate = NULL, const = NULL, flow)
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}

\item{from}{The name of the source compartment(s) (character vector or NULL for source compartments)}

\item{to}{The name of the destination compartment(s) (character vector or NULL for sink compartments)}

\item{...}{Unused, enforces \code{rate} and \code{const} to be specified as named arguments only, not positional}

\item{rate}{The flow rate (numeric, optional, mutually exclusive with \code{const})}

\item{const}{Rate constant for first-order flows (numeric, optional, mutually exclusive with \code{rate})}

\item{flow}{A \code{Flows} object. Constructed from the other inputs if not provided.}
}
\value{
The modified \code{CompartmentModel} object.
}
\description{
Flows represent mass/amount-preserving transitions between two compartments in a pharmacokinetic model.
}
\examples{
# Add a linear flow
compartment_model() |>
    add_flow(from = "A", to = "B", const = "k1")
# Add a nonlinear flow
compartment_model() |>
    add_flow(from = "A", to = "B", rate = "k2 * A / (A + Km)")
# Vectorized calls using naming abbreviation
compartment_model() |>
    add_flow("A", c("B", "C"), const = "k_to")
compartment_model() |>
    add_flow(c("A", "B"), c("B", "C"), const = "k_from_to")
}
