% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{add_meta}
\alias{add_meta}
\title{Add metadata (categorical attributes) to a \code{Physiology} object as name-value pairs.}
\usage{
add_meta(phys, ...)
}
\arguments{
\item{phys}{A \code{Physiology} object}

\item{...}{Name-value pairs.}
}
\value{
The updated \code{Physiology} object
}
\description{
Add metadata (categorical attributes) to a \code{Physiology} object as name-value pairs.
}
\examples{
physiology() |>
    add_meta(species = "rat", strain = "Wistar", sex = "male")
}
