% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{add_scalar}
\alias{add_scalar}
\title{Add a scalar parameter to a \code{Physiology} object}
\usage{
add_scalar(phys, name, value, unit = "", reference = "", assumption = "")
}
\arguments{
\item{phys}{A \code{Physiology} object}

\item{name}{Name of the parameter}

\item{value}{Numeric value}

\item{unit}{Character, unit of measurement}

\item{reference}{Source reference}

\item{assumption}{Assumptions made}
}
\value{
The updated \code{Physiology} object
}
\description{
Add a scalar parameter to a \code{Physiology} object
}
\examples{
physiology() |>
    add_scalar(name = "BW", value = 250)
}
