% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{add_tissue_param}
\alias{add_tissue_param}
\title{Add a tissue parameter (e.g. volume, flow) to a \code{Physiology} object}
\usage{
add_tissue_param(
  phys,
  tissue,
  name,
  value,
  unit = "",
  reference = "",
  assumption = ""
)
}
\arguments{
\item{phys}{A \code{Physiology} object}

\item{tissue}{Tissue name (e.g. "Liver")}

\item{name}{Parameter name (e.g. "V" or "Q")}

\item{value}{Numeric value}

\item{unit}{Character, unit of measurement}

\item{reference}{Source reference}

\item{assumption}{Assumptions made}
}
\value{
The updated \code{Physiology} object
}
\description{
Add a tissue parameter (e.g. volume, flow) to a \code{Physiology} object
}
