% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartments.R
\name{as.data.frame.Compartments}
\alias{as.data.frame.Compartments}
\title{Convert a \code{Compartments} object to a data frame}
\usage{
\method{as.data.frame}{Compartments}(x, ...)
}
\arguments{
\item{x}{A \code{Compartments} object}

\item{...}{Additional arguments (not used)}
}
\value{
A data frame representation of the \code{Compartments} object
}
\description{
Convert a \code{Compartments} object to a data frame
}
