% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dosing.R
\name{as.data.frame.Dosing}
\alias{as.data.frame.Dosing}
\title{Convert a \code{Dosing} object to a data frame}
\usage{
\method{as.data.frame}{Dosing}(x, ...)
}
\arguments{
\item{x}{A \code{Dosing} object}

\item{...}{ignored}
}
\value{
A data frame representation of the \code{Dosing} object
}
\description{
This method allows a \code{Dosing} object to be treated as a data frame for operations
like subsetting and combining, while preserving the \code{Dosing} class.
}
