% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flows.R
\name{as.data.frame.Flows}
\alias{as.data.frame.Flows}
\title{Convert a \code{Flows} object to a data frame}
\usage{
\method{as.data.frame}{Flows}(x, ...)
}
\arguments{
\item{x}{A \code{Flows} object}

\item{...}{Additional arguments (not used)}
}
\value{
A data frame representation of the \code{Flows} object
}
\description{
Convert a \code{Flows} object to a data frame
}
