% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecules.R
\name{as.data.frame.Molecules}
\alias{as.data.frame.Molecules}
\title{Convert a \code{Molecules} object to a data frame}
\usage{
\method{as.data.frame}{Molecules}(x, ...)
}
\arguments{
\item{x}{A \code{Molecules} object}

\item{...}{Additional arguments (not used)}
}
\value{
A data frame representation of the \code{Molecules} object
}
\description{
Convert a \code{Molecules} object to a data frame
}
