% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactions.R
\name{as.data.frame.Reactions}
\alias{as.data.frame.Reactions}
\title{Convert a \code{Reactions} object to a data frame}
\usage{
\method{as.data.frame}{Reactions}(x, ...)
}
\arguments{
\item{x}{A \code{Reactions} object}

\item{...}{Additional arguments (not used)}
}
\value{
A data frame representation of the \code{Reactions} object
}
\description{
Convert a \code{Reactions} object to a data frame
}
