% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transports.R
\name{as.data.frame.Transports}
\alias{as.data.frame.Transports}
\title{Convert a \code{Transports} object to a data frame}
\usage{
\method{as.data.frame}{Transports}(x, ...)
}
\arguments{
\item{x}{A \code{Transports} object}

\item{...}{Additional arguments (not used)}
}
\value{
A data frame representation of the \code{Transports} object
}
\description{
Convert a \code{Transports} object to a data frame
}
