% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flows.R
\name{as.list.Flows}
\alias{as.list.Flows}
\title{Convert a \code{Flows} object to a list of \code{Flow} objects.}
\usage{
\method{as.list}{Flows}(x, ...)
}
\arguments{
\item{x}{A \code{Flows} object}

\item{...}{Additional arguments (not used)}
}
\value{
A list of \code{Flow} objects, where each component is a list of the form \code{list(from, to, rate, const, type)}
}
\description{
\code{Flow} objects are light-weighted representations of individual flows,
where each flow is represented as a list with entries
\code{from}, \code{to}, \code{rate}, \code{const}, and \code{type}.
This function is useful for iterating over flows in a \code{Flows} object.
}
