% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecules.R
\name{as.list.Molecules}
\alias{as.list.Molecules}
\title{Convert a \code{Molecules} object to a list of lists, where each inner list represents a molecule with its properties}
\usage{
\method{as.list}{Molecules}(x, ...)
}
\arguments{
\item{x}{A \code{Molecules} object}

\item{...}{Additional arguments (not used)}
}
\value{
A list of lists, where each inner list represents a molecule with its properties
}
\description{
Convert a \code{Molecules} object to a list of lists, where each inner list represents a molecule with its properties
}
