% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactions.R
\name{as.list.Reactions}
\alias{as.list.Reactions}
\title{Convert a \code{Reactions} object to a list of lists, where each inner list represents a reaction with its properties}
\usage{
\method{as.list}{Reactions}(x, ...)
}
\arguments{
\item{x}{A \code{Reactions} object}

\item{...}{Additional arguments (not used)}
}
\value{
A list of lists, where each inner list represents a reaction with its properties
}
\description{
Convert a \code{Reactions} object to a list of lists, where each inner list represents a reaction with its properties
}
