% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transports.R
\name{as.list.Transports}
\alias{as.list.Transports}
\title{Convert a \code{Transports} object to a list of \code{Transport} objects (experimental).}
\usage{
\method{as.list}{Transports}(x, ...)
}
\arguments{
\item{x}{A \code{Transports} object}

\item{...}{Additional arguments (not used)}
}
\value{
A list of \code{Transport} objects, where each component is a list of the form \code{list(from, to, molec, rate, const, type)}
}
\description{
\code{Transport} objects are light-weighted representations of individual transports,
where each transport is represented as a list with entries
\code{from}, \code{to}, \code{molec}, \code{rate}, \code{const}, and \code{type}.
This function is useful for iterating over transports in a \code{Transports} object.
}
