% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationData.R
\name{as_observation_schedule}
\alias{as_observation_schedule}
\title{Extract the schedule from observation data}
\usage{
as_observation_schedule(x)
}
\arguments{
\item{x}{An \code{ObservationSchedule} or \code{ObservationData} object.}
}
\value{
An \code{ObservationSchedule} without the \code{value} column.
}
\description{
Extract the schedule from observation data
}
