% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drugs.R
\name{c.Drug}
\alias{c.Drug}
\alias{c.Drugs}
\title{Combine multiple Drug or Drugs objects into a Drugs object}
\usage{
\method{c}{Drug}(...)

\method{c}{Drugs}(...)
}
\arguments{
\item{...}{\code{Drug} or \code{Drugs} objects to combine}
}
\value{
A \code{Drugs} object containing the combined drugs.
}
\description{
This method allows you to combine multiple \code{Drug} or \code{Drugs} objects into a \code{Drugs} using the \code{c()} function.
}
