% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{c.Equations}
\alias{c.Equations}
\title{Combine multiple \code{Equations} objects}
\usage{
\method{c}{Equations}(...)
}
\arguments{
\item{...}{\code{Equations} objects to combine}
}
\value{
A combined \code{Equations} object
}
\description{
Combine multiple \code{Equations} objects
}
