% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiments.R
\name{c.Experiments}
\alias{c.Experiments}
\title{Combine collections of experiments}
\usage{
\method{c}{Experiments}(...)
}
\arguments{
\item{...}{\code{Experiments} objects to combine.}
}
\value{
An \code{Experiments} object, preserving order and names.
}
\description{
Combine collections of experiments
}
