% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flows.R
\name{c.Flow}
\alias{c.Flow}
\title{Combine multiple \code{Flow} objects into a \code{Flows} object}
\usage{
\method{c}{Flow}(...)
}
\arguments{
\item{...}{Multiple \code{Flow} objects to combine}
}
\value{
A combined \code{Flows} object
}
\description{
Combine multiple \code{Flow} objects into a \code{Flows} object
}
