% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flows.R
\name{c.Flows}
\alias{c.Flows}
\title{Combine multiple \code{Flows} objects into one}
\usage{
\method{c}{Flows}(...)
}
\arguments{
\item{...}{Multiple \code{Flows} objects to combine}
}
\value{
A combined \code{Flows} object
}
\description{
Combine multiple \code{Flows} objects into one
}
