% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecules.R
\name{c.Molecules}
\alias{c.Molecules}
\title{Combine multiple \code{Molecules} objects into one}
\usage{
\method{c}{Molecules}(...)
}
\arguments{
\item{...}{Multiple \code{Molecules} objects to combine}
}
\value{
A combined \code{Molecules} object
}
\description{
Combine multiple \code{Molecules} objects into one
}
