% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{c.Parameters}
\alias{c.Parameters}
\title{Combine multiple \code{Parameters} objects into one}
\usage{
\method{c}{Parameters}(...)
}
\arguments{
\item{...}{\code{Parameters} objects to combine}
}
\value{
A combined \code{Parameters} object
}
\description{
Combine multiple \code{Parameters} objects into one
}
