% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSets.R
\name{c.ParameterSets}
\alias{c.ParameterSets}
\title{Combine realized parameter-set collections}
\usage{
\method{c}{ParameterSets}(...)
}
\arguments{
\item{...}{\code{ParameterSets} objects to combine.}
}
\value{
A \code{ParameterSets} object, preserving order and names.
}
\description{
Combine realized parameter-set collections
}
