% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactions.R
\name{c.Reactions}
\alias{c.Reactions}
\title{Combine multiple \code{Reactions} objects into one}
\usage{
\method{c}{Reactions}(...)
}
\arguments{
\item{...}{Multiple \code{Reactions} objects to combine}
}
\value{
A combined \code{Reactions} object
}
\description{
Combine multiple \code{Reactions} objects into one
}
