% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transports.R
\name{c.Transport}
\alias{c.Transport}
\title{Combine multiple \code{Transport} objects into a \code{Transports} object}
\usage{
\method{c}{Transport}(...)
}
\arguments{
\item{...}{Multiple \code{Transport} objects to combine}
}
\value{
A combined \code{Transports} object
}
\description{
Combine multiple \code{Transport} objects into a \code{Transports} object
}
