% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transports.R
\name{c.Transports}
\alias{c.Transports}
\title{Combine multiple \code{Transports} objects into one}
\usage{
\method{c}{Transports}(...)
}
\arguments{
\item{...}{Multiple \code{Transports} objects to combine}
}
\value{
A combined \code{Transports} object
}
\description{
Combine multiple \code{Transports} objects into one
}
