% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dosing.R
\name{$.Dosing}
\alias{$.Dosing}
\title{Extract a column from a \code{Dosing} object}
\usage{
\method{$}{Dosing}(x, name)
}
\arguments{
\item{x}{A \code{Dosing} object}

\item{name}{Column name}
}
\value{
The requested column.
}
\description{
Amount and rate columns are returned as vectors when units are compatible and
as lists otherwise.
}
