% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{coef.EstimationResult}
\alias{coef.EstimationResult}
\title{Extract estimated coefficients}
\usage{
\method{coef}{EstimationResult}(object, ...)
}
\arguments{
\item{object}{An \code{EstimationResult}.}

\item{...}{Unused.}
}
\value{
Estimated coefficients as a \code{Parameters} object.
}
\description{
Extract estimated coefficients
}
