% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartment.R
\name{Compartment}
\alias{Compartment}
\title{Compartment Class}
\description{
Represents a model compartment with a name and an initial amount.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Compartment-new}{\code{Compartment$new()}}
\item \href{#method-Compartment-clone}{\code{Compartment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compartment-new"></a>}}
\if{latex}{\out{\hypertarget{method-Compartment-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$new(name, initial = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compartment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Compartment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
