% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{compartment_model}
\alias{compartment_model}
\title{Create a new \code{CompartmentModel} object.}
\usage{
compartment_model()
}
\value{
A new \code{CompartmentModel} object.
}
\description{
A compartment model consists of compartments, molecules, transports between compartments,
reactions between molecules, equations defining auxiliary variables,
observables defined as functions of the states and parameters, parameters, and dosing events.
This function initializes an empty model that can then be built up by adding these components using the \verb{add_*} functions.
}
\seealso{
\code{\link[=add_compartment]{add_compartment()}}, \code{\link[=add_molecule]{add_molecule()}}, \code{\link[=add_transport]{add_transport()}}, \code{\link[=add_reaction]{add_reaction()}}, \code{\link[=add_equation]{add_equation()}},
\code{\link[=add_observable]{add_observable()}}, \code{\link[=add_parameter]{add_parameter()}}, \code{\link[=add_dosing]{add_dosing()}} for building up a \code{CompartmentModel} object,
and \code{\link[=to_ode]{to_ode()}}, \code{\link[=to_analytical]{to_analytical()}} for exporting a \code{CompartmentModel} to ODE or analytical solution format.
}
