% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{compartment_names}
\alias{compartment_names}
\title{Get compartment names of a \code{CompartmentModel}.}
\usage{
compartment_names(model)
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}
}
\value{
A character vector of compartment names.
}
\description{
Get compartment names of a \code{CompartmentModel}.
}
