% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartment.R
\name{CompartmentList}
\alias{CompartmentList}
\title{Construct one or more Compartment objects}
\usage{
CompartmentList(name, initial = 0)
}
\arguments{
\item{name}{Compartment name(s), character scalar or vector}

\item{initial}{Initial amounts for the compartments (numeric scalar or vector, default = 0)}
}
\value{
A list of \code{Compartment} objects
}
\description{
\code{CompartmentList()} is a wrapper function to construct multiple \code{Compartment} objects at once.
}
