% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel-methods.R
\name{CompartmentModel$print}
\alias{CompartmentModel$print}
\title{Print method for CompartmentModel class}
\arguments{
\item{...}{ignored}
}
\value{
Invisibly returns \code{x}.
}
\description{
Pretty-prints a \code{CompartmentModel} object.
}
\examples{
M <- multiCompModel()
print(M)
}
\seealso{
\link{CompartmentModel}
}
