% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel-class.R
\name{CompartmentModel}
\alias{CompartmentModel}
\title{\code{CompartmentModel} class}
\description{
\code{CompartmentModel} class

\code{CompartmentModel} class
}
\details{
Represents a compartmental model.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{compartments}}{A list of \code{Compartment} objects}

\item{\code{reactions}}{A list of \code{Reaction} objects}

\item{\code{observables}}{A list of \code{Observable} objects}

\item{\code{doses}}{List of dosing objects. Includes bolus doses into model
compartments, as well as bolus-to-bag doses used to implement infusions.}

\item{\code{infusionEvents}}{Data frame of infusion rate adjustment events
(start and stop of zero-order input). Generated automatically when
infusions are defined.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CompartmentModel-new}{\code{CompartmentModel$new()}}
\item \href{#method-CompartmentModel-print}{\code{CompartmentModel$print()}}
\item \href{#method-CompartmentModel-getInitialState}{\code{CompartmentModel$getInitialState()}}
\item \href{#method-CompartmentModel-getStateNames}{\code{CompartmentModel$getStateNames()}}
\item \href{#method-CompartmentModel-addCompartment}{\code{CompartmentModel$addCompartment()}}
\item \href{#method-CompartmentModel-addReaction}{\code{CompartmentModel$addReaction()}}
\item \href{#method-CompartmentModel-addObservable}{\code{CompartmentModel$addObservable()}}
\item \href{#method-CompartmentModel-addDosing}{\code{CompartmentModel$addDosing()}}
\item \href{#method-CompartmentModel-dosing_to_events}{\code{CompartmentModel$dosing_to_events()}}
\item \href{#method-CompartmentModel-toAnalytical}{\code{CompartmentModel$toAnalytical()}}
\item \href{#method-CompartmentModel-toODE}{\code{CompartmentModel$toODE()}}
\item \href{#method-CompartmentModel-clone}{\code{CompartmentModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new \code{CompartmentModel} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-print"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-getInitialState"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-getInitialState}{}}}
\subsection{Method \code{getInitialState()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$getInitialState(named = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-getStateNames"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-getStateNames}{}}}
\subsection{Method \code{getStateNames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$getStateNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-addCompartment"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-addCompartment}{}}}
\subsection{Method \code{addCompartment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$addCompartment(name, initial = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-addReaction"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-addReaction}{}}}
\subsection{Method \code{addReaction()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$addReaction(from, to, rate)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-addObservable"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-addObservable}{}}}
\subsection{Method \code{addObservable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$addObservable(name, expr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-addDosing"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-addDosing}{}}}
\subsection{Method \code{addDosing()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$addDosing(dose)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-dosing_to_events"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-dosing_to_events}{}}}
\subsection{Method \code{dosing_to_events()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$dosing_to_events()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-toAnalytical"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-toAnalytical}{}}}
\subsection{Method \code{toAnalytical()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$toAnalytical(paramValues = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-toODE"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-toODE}{}}}
\subsection{Method \code{toODE()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$toODE(paramValues = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
