% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic-lumping-algebra.R
\name{.assemble_linear_expr}
\alias{.assemble_linear_expr}
\title{Assemble the lumping equations for a strongly connected component (SCC)}
\usage{
.assemble_linear_expr(scc, reactions)
}
\arguments{
\item{scc}{A character vector, the strongly connected component (SCC) to be lumped}

\item{reactions}{A list of \code{Reaction} objects}
}
\value{
A list with entries \code{A}, \code{b}
}
\description{
Assemble the lumping equations for a strongly connected component (SCC)
}
