% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic-lumping-algebra.R
\name{.classify_flows}
\alias{.classify_flows}
\title{Symbolic lumping of a linear \code{CompartmentModel} based on its graph structure}
\usage{
.classify_flows(flows, scc)
}
\arguments{
\item{flows}{A \code{Flows} object}

\item{scc}{A character vector, the strongly connected component (SCC) to be lumped}
}
\value{
A list with entries \code{internal}, \code{incoming}, \code{outgoing}, \code{elimination}, each a list of \code{Reaction} objects
}
\description{
Symbolic lumping of a linear \code{CompartmentModel} based on its graph structure
}
