% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{.dosing_to_events}
\alias{.dosing_to_events}
\title{Generate events data.frame for \code{deSolve} from stored dosing.}
\usage{
.dosing_to_events(model)
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}
}
\value{
A list with a single element \code{data}, which is a data.frame with columns \code{var}, \code{time}, \code{value}, and \code{method} (add or replace).
}
\description{
This unifies bolus and infusion dosing into a single events data.frame, which is then used in
the \code{toODE} method to generate the final events list for \code{deSolve}.
}
