% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic-lumping-graph.R
\name{.make_graph}
\alias{.make_graph}
\title{Construct a directed graph encoding the topology of a linear \code{CompartmentModel}.}
\usage{
.make_graph(model, refstate)
}
\arguments{
\item{model}{A \code{CompartmentModel} object}

\item{`refstate`}{A string, corresponding to the state used as reference.}
}
\value{
A length two named list with entries \code{edges} (a data frame with
columns \code{from} and \code{to}) and \code{nodes} (a character vector)
}
\description{
Construct a directed graph encoding the topology of a linear \code{CompartmentModel}.
}
