% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumping.R
\name{.rewrite_rate}
\alias{.rewrite_rate}
\title{Helper function to replace reaction rates during (un-)lumping by AST traversal}
\usage{
.rewrite_rate(expr, grp, VKorig)
}
\description{
Helper function to replace reaction rates during (un-)lumping by AST traversal
}
