% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic-lumping-algebra.R
\name{.simplify_expr}
\alias{.simplify_expr}
\title{Simplify a symbolic expression by cancelling double negatives}
\usage{
.simplify_expr(expr)
}
\arguments{
\item{expr}{A quoted call}
}
\value{
A simplified expression
}
\description{
Simplify a symbolic expression by cancelling double negatives
}
