% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic-lumping-algebra.R
\name{.solve_linear_expr}
\alias{.solve_linear_expr}
\title{Solve a linear system of equations Ax+b=0 with symbolic coefficients}
\usage{
.solve_linear_expr(A, b)
}
\arguments{
\item{A}{A square list matrix of expressions}

\item{b}{A list of right-hand side expressions}
}
\value{
A named list of solution expressions
}
\description{
Solve a linear system of equations Ax+b=0 with symbolic coefficients
}
