% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drugs.R
\name{drug}
\alias{drug}
\title{Class representing drug data}
\usage{
drug(..., data = NULL)
}
\arguments{
\item{...}{Named arguments representing drug parameters (possibly with units).}

\item{data}{A data frame with the following columns: "Parameter", "Species", "Value", "Unit", "Reference", "Assumption".
This can be used as an alternative to specifying parameters via \code{...}.}
}
\description{
This class is used to store drug-specific data, such as physicochemical properties and ADME parameters, that can be used in PBPK models.
It is designed to be flexible and extensible, allowing users to add new properties as needed.
}
