% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{empirical_pk_receptor}
\alias{empirical_pk_receptor}
\title{Empirical PK model relative to number of receptors}
\usage{
empirical_pk_receptor()
}
\value{
A CompartmentModel object
}
\description{
Empirical PK model relative to number of receptors
}
