% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartments.R
\name{empty_compartment}
\alias{empty_compartment}
\title{Create an empty \code{Compartments} object}
\usage{
empty_compartment()
}
\value{
An empty \code{Compartments} object
}
\description{
Create an empty \code{Compartments} object
}
