% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dosing.R
\name{empty_dosing}
\alias{empty_dosing}
\title{Create an empty \code{Dosing} object}
\usage{
empty_dosing()
}
\value{
An empty \code{Dosing} object with zero rows
}
\description{
Create an empty \code{Dosing} object
}
