% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{empty_equation}
\alias{empty_equation}
\title{Create an empty \code{Equations} object}
\usage{
empty_equation()
}
\value{
An empty \code{Equations} object
}
\description{
Create an empty \code{Equations} object
}
