% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observable.R
\name{empty_observable}
\alias{empty_observable}
\title{Create an empty \code{Observables} object}
\usage{
empty_observable()
}
\value{
An empty \code{Observables} object
}
\description{
Create an empty \code{Observables} object
}
