% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{estimate}
\alias{estimate}
\alias{estimate.EstimationProblem}
\title{Estimate model parameters}
\usage{
estimate(object, ...)

\method{estimate}{EstimationProblem}(object, engine = optim_engine(), dimensions = NULL, ...)
}
\arguments{
\item{object}{An \code{EstimationProblem}.}

\item{...}{Additional method arguments.}

\item{engine}{An estimation engine, currently \code{\link[=optim_engine]{optim_engine()}}.}

\item{dimensions}{Optional solver-facing unit dimensions passed to simulation.}
}
\description{
Estimate model parameters
}
