% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{estimated_parameters}
\alias{estimated_parameters}
\title{Collect estimated parameter specifications}
\usage{
estimated_parameters(...)
}
\arguments{
\item{...}{Named objects created by \code{\link[=parameter_estimate]{parameter_estimate()}}.}
}
\value{
An \code{EstimatedParameters} object.
}
\description{
Collect estimated parameter specifications
}
