% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic-lumping.R
\name{get_lumping_conditions}
\alias{get_lumping_conditions}
\title{Get symbolic lumping conditions based on a reference state}
\usage{
get_lumping_conditions(
  M,
  refstate,
  maxdegree = 2,
  simplify = c("none", "Ryacas")
)
}
\arguments{
\item{M}{A \code{CompartmentModel} object}

\item{refstate}{A string specifying the reference state}

\item{maxdegree}{An integer specifying the maximum degree for which to carry out symbolic substitution.
Values > 2 are currently unsupported and will throw an error.}

\item{simplify}{Method for simplifying the resulting expressions.
Install Ryacas package to use this feature.}
}
\value{
A named list of expressions representing the lumping conditions. Names are compartment names and
expressions are in terms of the reference state and rate constants.
}
\description{
Get symbolic lumping conditions based on a reference state
}
\examples{
# Example: Pilari/Huisinga (2010) lumping conditions for PBPK 12-compartment model
M <- sMD_PBPK_12CMT_wellstirred()
res <- get_lumping_conditions(M, refstate = "ven", simplify = "Ryacas")
}
