% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{getopt_compphysiol}
\alias{getopt_compphysiol}
\title{Get package options}
\usage{
getopt_compphysiol(name = NULL)
}
\arguments{
\item{name}{Option name, or NULL for all options}
}
\description{
Get package options
}
