% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{initials}
\alias{initials}
\title{Extract the initial states (with names) from a \code{CompartmentModel} object}
\usage{
initials(model, type = c("a[] or c[]", "c[] or a[]", "a[] only", "c[] only"))
}
\arguments{
\item{model}{A \code{CompartmentModel} object}

\item{type}{The type of initials to extract: the default \code{"a[] or c[]"} means amounts are attempted first,
but concentrations given if amounts cannot be calculated. \code{"c[] or a[]"} does the opposite,
while \code{"a[] only"} and \code{"c[] only"} will return an error if the respective type cannot be calculated.
If all compartments have volumes (fixed or parametrized), then both types of initials can be calculated
and the \code{type} argument only determines which one is returned. For compartments that have no defined volume,
only the type of initial specified in the molecule definition can be calculated.}
}
\value{
A named list of state initial values, where the names are in the format
\code{"a[molec,cmt]"} for amount states or \code{"c[molec,cmt]"} for concentration states.
}
\description{
Extract the initial states (with names) from a \code{CompartmentModel} object
}
