% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionization.R
\name{ionized_fractions}
\alias{ionized_fractions}
\title{Ionized fractions for a small molecule drug}
\usage{
ionized_fractions(
  pH,
  pKa_ani = NULL,
  pKa_cat = NULL,
  Kz = 0,
  respectThermodynamics = TRUE
)
}
\arguments{
\item{pH}{Numeric vector of (environmental) pH value(s)}

\item{pKa_ani}{Anionic pKa value(s), i.e. leading to a negative charge}

\item{pKa_cat}{Cationic pKa value(s), i.e. leading to a positive charge}

\item{Kz}{Tautomeric constant for ampholytes (substances with non-empty
\code{pKa_ani} and \code{pKa_cat}), which defines the ratio of zwitterionic to
neutral fractions, i.e. \code{Kz = fz/fn} (the value \code{Kz = +Inf} is allowed
and means that the shift from anionic to cationic species occurs
exclusively via the zwitter ion). For ampholytes, it is required
that \code{pKa_cat <= pKa_ani}, which can be derived from theory based on
electrostatic interactions.}

\item{respectThermodynamics}{A boolean that allows to
specify whether pKa values of ampholytes must be ordered in accordance
with thermodynamics (every anionic pKa larger than any cationic pKa)
(default: \code{TRUE}). This option is mainly provided to be able to reproduce
Rodgers et al.'s results, it is not recommended to change the default.}
}
\value{
A list with fields \code{n},\code{ani},\code{cat},\code{z} for neutral, anionic, cationic,
and zwitterionic fractions.
}
\description{
Ionized fractions for a small molecule drug
}
\examples{
# neutral drug
ionized_fractions(pH = 7.4)
# monoprotic acid
ionized_fractions(pH = 7.4, pKa_ani = 5)
# monoprotic base
ionized_fractions(pH = 7.4, pKa_cat = 10)
# diprotic acid
ionized_fractions(pH = 7.4, pKa_ani = c(5,7))
# diprotic base
ionized_fractions(pH = 7.4, pKa_cat = c(8,10))
# ordinary ampholyte
ionized_fractions(pH = 7.4, pKa_ani = 10, pKa_cat = 5)
# purely zwitterionic ampholyte
ionized_fractions(pH = 7.4, pKa_ani = 10, pKa_cat = 5, Kz = Inf)

}
