% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dosing.R
\name{is_bolus}
\alias{is_bolus}
\title{Check which dosing events are boluses.}
\usage{
is_bolus(dose)
}
\arguments{
\item{dose}{A \code{Dosing} object}
}
\value{
A logical vector in which the i-th entry is \code{TRUE} if the i-th dosing is a bolus and \code{FALSE} otherwise
}
\description{
Check which dosing events are boluses.
}
