% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dosing.R
\name{is_infusion}
\alias{is_infusion}
\title{Check which dosing events are infusions.}
\usage{
is_infusion(dose)
}
\arguments{
\item{dose}{A \code{Dosing} object}
}
\value{
A logical vector in which the i-th entry is \code{TRUE} if the i-th dosing is an infusion and \code{FALSE} otherwise
}
\description{
Check which dosing events are infusions.
}
