% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flows.R
\name{length.Flows}
\alias{length.Flows}
\title{Length of a \code{Flows} object}
\usage{
\method{length}{Flows}(x)
}
\arguments{
\item{x}{A \code{Flows} object}
}
\value{
The number of flows in the object
}
\description{
Length of a \code{Flows} object
}
