% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecules.R
\name{length.Molecules}
\alias{length.Molecules}
\title{Length of a \code{Molecules} object}
\usage{
\method{length}{Molecules}(x, ...)
}
\arguments{
\item{x}{A \code{Molecules} object}

\item{...}{Additional arguments (not used)}
}
\value{
The number of molecules in the object
}
\description{
Length of a \code{Molecules} object
}
