% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactions.R
\name{length.Reactions}
\alias{length.Reactions}
\title{Length of a \code{Reactions} object}
\usage{
\method{length}{Reactions}(x)
}
\arguments{
\item{x}{A \code{Reactions} object}
}
\value{
The number of reactions in the object
}
\description{
Length of a \code{Reactions} object
}
