% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transports.R
\name{length.Transports}
\alias{length.Transports}
\title{Length of a \code{Transports} object}
\usage{
\method{length}{Transports}(x)
}
\arguments{
\item{x}{A \code{Transports} object}
}
\value{
The number of transports in the object
}
\description{
Length of a \code{Transports} object
}
