% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drugs.R
\name{loaddrugdata}
\alias{loaddrugdata}
\title{Load a drug from the drug database}
\usage{
loaddrugdata(name = "", species = NULL)
}
\arguments{
\item{name}{The name of the drug to load.}

\item{species}{Character string, for which species to load species-specific parameters (default: all species).}
}
\value{
A \code{Drug} object containing all parameters for the specified drug and species.
}
\description{
Load a drug from the drug database
}
