% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{loadphysiology}
\alias{loadphysiology}
\title{Load a physiology from the physiology database}
\usage{
loadphysiology(name)
}
\arguments{
\item{name}{The name of the physiology to load (e.g. "rat250" or "human35m")}
}
\value{
A \code{Physiology} object containing the parameters and metadata for the specified physiology.
}
\description{
Load a physiology from the physiology database
}
