% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{make_depot}
\alias{make_depot}
\title{Make auxiliary structures in a \code{CompartmentModel} object for handling continuous inputs}
\usage{
make_depot(model)
}
\arguments{
\item{model}{A \code{CompartmentModel} object}
}
\value{
An updated \code{CompartmentModel} object with the auxiliary compartments and events added.
}
\description{
This function adds auxiliary compartments and events to the model to handle continuous inputs:
\itemize{
\item \verb{Depot_\{molec\}_\{cmt\}} and \verb{ReleaseRate_\{molec\}_\{cmt\}} compartments
(both with \code{NA} volumes and for each molecule / target compartment)
\item transports from the depot to its target compartment, at rate \verb{a[ReleaseRate_\{molec\}_\{cmt\}]}
\item 3 (instantaneous) dosing events per infusion (fill depot at infusion start,
increase release rate at infusion start, decrease release rate at infusion end)
}
}
