% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-units.R
\name{model_unit_registry}
\alias{model_unit_registry}
\title{Inspect registered model units}
\usage{
model_unit_registry()
}
\value{
A data frame with registered unit symbols and definitions.
}
\description{
Inspect registered model units
}
