% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecules.R
\name{names.Molecules}
\alias{names.Molecules}
\title{Names of molecules in a \code{Molecules} object}
\usage{
\method{names}{Molecules}(x, ...)
}
\arguments{
\item{x}{A \code{Molecules} object}

\item{...}{Additional arguments (not used)}
}
\value{
A character vector of molecule names
}
\description{
Names of molecules in a \code{Molecules} object
}
