% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observables.R
\name{names.Observables}
\alias{names.Observables}
\title{Names of observables in an \code{Observables} object}
\usage{
\method{names}{Observables}(x)
}
\arguments{
\item{x}{An \code{Observables} object}
}
\value{
The names of the observables
}
\description{
Names of observables in an \code{Observables} object
}
