% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observable.R
\name{Observable}
\alias{Observable}
\title{Observable Class}
\description{
Represents a model observable
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Observable-new}{\code{Observable$new()}}
\item \href{#method-Observable-clone}{\code{Observable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Observable-new"></a>}}
\if{latex}{\out{\hypertarget{method-Observable-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observable$new(name, expr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Observable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Observable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
