% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observable.R
\name{ObservableList}
\alias{ObservableList}
\title{Construct one or more Observable objects}
\usage{
ObservableList(name, expr)
}
\arguments{
\item{name}{Observable name(s), character scalar or vector}

\item{expr}{Observable expression(s), (lists of) expressions or functions, or character scalar or vector}
}
\value{
A list of \code{Observable} objects.
}
\description{
\code{ObservableList()} is a wrapper function to construct multiple \code{Observable} objects at once.
It accepts vectorized arguments of identical length.
}
\examples{
ObservableList(c("Cpla","Cblo"), expr = c("Central/Vcentral/BP", "Central/Vcentral"))

}
