% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{observation_errors}
\alias{observation_errors}
\alias{additive_error}
\alias{proportional_error}
\alias{combined_error}
\alias{lognormal_error}
\title{Construct residual-error specifications}
\usage{
additive_error(sigma = "sigma")

proportional_error(sigma = "sigma")

combined_error(additive = "sigma_add", proportional = "sigma_prop")

lognormal_error(sigma = "sigma")
}
\arguments{
\item{sigma}{Name of the residual standard-deviation parameter.}

\item{additive}{Name of the additive standard-deviation parameter.}

\item{proportional}{Name of the proportional standard-deviation parameter.}
}
\value{
An \code{ObservationError} object of the corresponding subclass.
}
\description{
Construct residual-error specifications
}
