% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{optim_backend}
\alias{optim_backend}
\title{Configure the stats optim estimation backend}
\usage{
optim_backend(method = "L-BFGS-B", control = list())
}
\arguments{
\item{method}{Optimization method passed to \code{\link[stats:optim]{stats::optim()}}.}

\item{control}{Control list passed to \code{\link[stats:optim]{stats::optim()}}.}
}
\value{
An \code{OptimBackend} object.
}
\description{
Configure the stats optim estimation backend
}
