% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{param}
\alias{param}
\title{Generic functions for accessing database parameters}
\usage{
param(x, name, ...)
}
\arguments{
\item{x}{The object to access}

\item{name}{The name of the parameter to access}

\item{...}{Additional arguments to be passed to specific methods}
}
\value{
A \code{Parameters} object containing the queried data.
}
\description{
Generic functions for accessing database parameters
}
