% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{parameter}
\alias{parameter}
\title{Get numeric parameter or metadata from a \code{Physiology} object by name (and optionally context)}
\usage{
parameter(phys, name, context = "scalar")
}
\arguments{
\item{phys}{A \code{Physiology} object}

\item{name}{Name of the parameter or metadata entry to be queried}

\item{context}{A string, defaulting to \code{"scalar"}}
}
\value{
The value of the parameter or metadata, or \code{NA} if not found
}
\description{
Get numeric parameter or metadata from a \code{Physiology} object by name (and optionally context)
}
\examples{
P <- physiology() |>
    add_scalar(name = "BW", value = 250)
parameter(P, "BW") # 250
}
