% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{physiology}
\alias{physiology}
\title{Create a new \code{Physiology} object}
\usage{
physiology(params = NULL, meta = NULL)
}
\arguments{
\item{params}{A data frame with physiology parameters}

\item{meta}{A list of metadata (e.g. species, sex)}
}
\value{
A new \code{Physiology} object
}
\description{
Create a new \code{Physiology} object
}
