% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessOdeModel.R
\name{print.AnalyticalModel}
\alias{print.AnalyticalModel}
\title{Print method for \code{AnalyticalModel} class}
\usage{
\method{print}{AnalyticalModel}(x, ...)
}
\arguments{
\item{x}{An \code{AnalyticalModel} object.}

\item{...}{ignored}
}
\value{
The \code{AnalyticalModel} object (invisibly).
}
\description{
Pretty-prints an \code{AnalyticalModel} object using state indices in the linear
system and DSL state names in the state list.
}
