% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{print.CompartmentModel}
\alias{print.CompartmentModel}
\title{Print method for CompartmentModel class
Pretty-prints a \code{CompartmentModel} object.}
\usage{
\method{print}{CompartmentModel}(x, ...)
}
\arguments{
\item{x}{A \code{CompartmentModel} object.}

\item{...}{ignored}
}
\value{
The \code{CompartmentModel} object (invisibly).
}
\description{
Print method for CompartmentModel class
Pretty-prints a \code{CompartmentModel} object.
}
