% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompiledOdeModel.R
\name{print.CompiledOdeModel}
\alias{print.CompiledOdeModel}
\title{Print method for \code{CompiledOdeModel} class}
\usage{
\method{print}{CompiledOdeModel}(x, ...)
}
\arguments{
\item{x}{A \code{CompiledOdeModel} object.}

\item{...}{ignored}
}
\value{
The \code{CompiledOdeModel} object (invisibly).
}
\description{
Pretty-prints the compiled model status, compact ODE model shape, and frozen
parameter interface.
}
