% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dosing.R
\name{print.Dosing}
\alias{print.Dosing}
\title{Print a \code{Dosing} object}
\usage{
\method{print}{Dosing}(x, ...)
}
\arguments{
\item{x}{A \code{Dosing} object}

\item{...}{ignored}
}
\value{
The \code{Dosing} object (invisibly)
}
\description{
Pretty-prints a \code{Dosing} object.
}
