% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drugs.R
\name{print.Drug}
\alias{print.Drug}
\title{Print method for Drug objects}
\usage{
\method{print}{Drug}(x, meta = FALSE, string = FALSE)
}
\arguments{
\item{x}{An object of class \code{Drug}.}

\item{meta}{Logical indicating whether to print parameter metadata as well (default: \code{FALSE}).}

\item{string}{Logical indicating whether to return a short string representation instead of printing (default: \code{FALSE}).}
}
\description{
Print method for Drug objects
}
