% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drugs.R
\name{print.Drugs}
\alias{print.Drugs}
\title{Print method for Drugs objects}
\usage{
\method{print}{Drugs}(x, meta = FALSE)
}
\arguments{
\item{x}{An object of class \code{Drugs}.}

\item{meta}{Logical indicating whether to print parameter metadata as well (default: \code{FALSE}).}
}
\description{
Print method for Drugs objects
}
