% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{print.EstimationProblem}
\alias{print.EstimationProblem}
\title{Print an estimation problem}
\usage{
\method{print}{EstimationProblem}(x, ...)
}
\arguments{
\item{x}{An \code{EstimationProblem} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an estimation problem
}
