% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{print.EstimationResult}
\alias{print.EstimationResult}
\title{Print an estimation result}
\usage{
\method{print}{EstimationResult}(x, ...)
}
\arguments{
\item{x}{An \code{EstimationResult} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an estimation result
}
