% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\name{print.Experiment}
\alias{print.Experiment}
\title{Print an experiment}
\usage{
\method{print}{Experiment}(x, ...)
}
\arguments{
\item{x}{An \code{Experiment} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an experiment
}
