% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiments.R
\name{print.Experiments}
\alias{print.Experiments}
\title{Print a collection of experiments}
\usage{
\method{print}{Experiments}(x, ...)
}
\arguments{
\item{x}{An \code{Experiments} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a collection of experiments
}
