% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flows.R
\name{print.Flow}
\alias{print.Flow}
\title{Print a \code{Flow} object}
\usage{
\method{print}{Flow}(x, ...)
}
\arguments{
\item{x}{A \code{Flow} object}

\item{...}{Additional arguments (not used)}
}
\value{
The \code{Flow} object (invisible)
}
\description{
Print a \code{Flow} object
}
