% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flows.R
\name{print.Flows}
\alias{print.Flows}
\title{Print a \code{Flows} object}
\usage{
\method{print}{Flows}(x, ...)
}
\arguments{
\item{x}{A \code{Flows} object}

\item{...}{Additional arguments (not used)}
}
\value{
The \code{Flows} object (invisible)
}
\description{
Print a \code{Flows} object
}
