% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecules.R
\name{print.Molecules}
\alias{print.Molecules}
\title{Print a \code{Molecules} object}
\usage{
\method{print}{Molecules}(x, ...)
}
\arguments{
\item{x}{A \code{Molecules} object}

\item{...}{Additional arguments (not used)}
}
\value{
The \code{Molecules} object (invisible)
}
\description{
Print a \code{Molecules} object
}
