% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observables.R
\name{print.Observables}
\alias{print.Observables}
\title{Print an \code{Observables} object}
\usage{
\method{print}{Observables}(x, ...)
}
\arguments{
\item{x}{An \code{Observables} object}

\item{...}{Additional arguments (not used)}
}
\value{
The \code{Observables} object (invisible)
}
\description{
Print an \code{Observables} object
}
