% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{print.ParameterEstimate}
\alias{print.ParameterEstimate}
\title{Print an estimated-parameter specification}
\usage{
\method{print}{ParameterEstimate}(x, ...)
}
\arguments{
\item{x}{A \code{ParameterEstimate} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an estimated-parameter specification
}
