% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{print.Parameters}
\alias{print.Parameters}
\title{Print a \code{Parameters} object}
\usage{
\method{print}{Parameters}(x, ...)
}
\arguments{
\item{x}{An \code{Parameters} object}

\item{...}{Additional arguments (not used)}
}
\value{
The \code{Parameters} object (invisible)
}
\description{
Print a \code{Parameters} object
}
