% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{print.ParameterSpec}
\alias{print.ParameterSpec}
\title{Print an estimated-parameter specification}
\usage{
\method{print}{ParameterSpec}(x, ...)
}
\arguments{
\item{x}{A \code{ParameterSpec} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an estimated-parameter specification
}
