% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{print.ParameterSpecs}
\alias{print.ParameterSpecs}
\title{Print parameter-estimation specifications}
\usage{
\method{print}{ParameterSpecs}(x, ...)
}
\arguments{
\item{x}{A \code{ParameterSpecs} collection.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print parameter-estimation specifications
}
