% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{print.Physiology}
\alias{print.Physiology}
\title{Print method for \code{Physiology} objects, showing metadata and a summary of parameters.}
\usage{
\method{print}{Physiology}(x, ...)
}
\arguments{
\item{x}{A \code{Physiology} object}

\item{...}{Additional arguments (not used)}
}
\value{
The \code{Physiology} object (invisible)
}
\description{
Print method for \code{Physiology} objects, showing metadata and a summary of parameters.
}
