% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessOdeModel.R
\name{print.ProcessModel}
\alias{print.ProcessModel}
\title{Print method for \code{ProcessModel} class}
\usage{
\method{print}{ProcessModel}(x, ...)
}
\arguments{
\item{x}{A \code{ProcessModel} object.}

\item{...}{ignored}
}
\value{
The \code{ProcessModel} object (invisibly).
}
\description{
Pretty-prints a \code{ProcessModel} object using DSL state names.
}
