% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{print.SimulationResult}
\alias{print.SimulationResult}
\title{Print a simulation result}
\usage{
\method{print}{SimulationResult}(x, ...)
}
\arguments{
\item{x}{A \code{SimulationResult} object.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
The \code{SimulationResult} object, invisibly.
}
\description{
Print a simulation result
}
