% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{print.StatisticalModel}
\alias{print.StatisticalModel}
\title{Print a statistical model}
\usage{
\method{print}{StatisticalModel}(x, ...)
}
\arguments{
\item{x}{A \code{StatisticalModel} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a statistical model
}
