% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessOdeModel.R
\name{print.StochasticModel}
\alias{print.StochasticModel}
\title{Print method for \code{StochasticModel} class}
\usage{
\method{print}{StochasticModel}(x, ...)
}
\arguments{
\item{x}{A \code{StochasticModel} object.}

\item{...}{ignored}
}
\value{
The \code{StochasticModel} object (invisibly).
}
\description{
Pretty-prints a \code{StochasticModel} object using DSL state names.
}
