% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transports.R
\name{print.Transport}
\alias{print.Transport}
\title{Print a \code{Transport} object}
\usage{
\method{print}{Transport}(x, ...)
}
\arguments{
\item{x}{A \code{Transport} object}

\item{...}{Additional arguments (not used)}
}
\value{
The \code{Transport} object (invisible)
}
\description{
Print a \code{Transport} object
}
