% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transports.R
\name{print.Transports}
\alias{print.Transports}
\title{Print a \code{Transports} object}
\usage{
\method{print}{Transports}(x, ...)
}
\arguments{
\item{x}{A \code{Transports} object}

\item{...}{Additional arguments (not used)}
}
\value{
The \code{Transports} object (invisible)
}
\description{
Print a \code{Transports} object
}
