% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reaction.R
\name{Reaction}
\alias{Reaction}
\title{Reaction Class}
\description{
Represents a model reaction.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Reaction-new}{\code{Reaction$new()}}
\item \href{#method-Reaction-print}{\code{Reaction$print()}}
\item \href{#method-Reaction-clone}{\code{Reaction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reaction-new"></a>}}
\if{latex}{\out{\hypertarget{method-Reaction-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reaction$new(from, to, rate)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reaction-print"></a>}}
\if{latex}{\out{\hypertarget{method-Reaction-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reaction$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reaction-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Reaction-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reaction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
