% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{receptor_system}
\alias{receptor_system}
\title{Receptor binding and internalization module}
\usage{
receptor_system(ligands = "L", dynamic = TRUE)
}
\arguments{
\item{ligands}{Character vector of ligand names (default: "L")}

\item{dynamic}{Whether to include ligand dynamics (default: TRUE). If FALSE, ligands
are treated as a (constant) parameter instead of a compartment.}
}
\value{
A CompartmentModel object
}
\description{
TODO: parametrize initial condition (initialize at steady-state -- ligand-free??)
}
