% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue-partitioning.R
\name{rodgersrowland}
\alias{rodgersrowland}
\title{Rodgers and Rowland's tissue partition prediction method}
\usage{
rodgersrowland(
  phys,
  drug,
  organs,
  plasmaWaterFraction = 0.93,
  fupIncludesLipids = TRUE,
  treatNegativeBindingAsZero = FALSE,
  respectThermodynamics = TRUE
)
}
\arguments{
\item{phys}{An object of class \code{Physiology} which defines the following
options for each tissue in argument \code{organs}:
\describe{
\item{TO DO}
}}

\item{drug}{An object of class \code{DrugData}.}

\item{organs}{A character array of valid organ (compartment) names
(e.g. \code{"adi"},\code{"bon"},...).}

\item{plasmaWaterFraction}{A value between 0 and 1 (default: \code{0.93})}

\item{fupIncludesLipids}{A boolean (default: \code{TRUE})}

\item{treatNegativeBindingAsZero}{A boolean (default: \code{FALSE})}

\item{respectThermodynamics}{A boolean (default: \code{TRUE})}
}
\description{
Function \code{rodgersrowland} predicts tissue-to-unbound plasma partition
coefficients using the method described in Rodgers et al. (2005,2006).
Rodgers/Rowland
Input PHYS is a Physiology object, DRUG a DrugData object and ORGANS a
cellstr with the compartment names.
Output is a struct that can be appended to the initialized model.
}
\details{
References:
\itemize{
\item Rodgers et al. (2005), DOI: 10.1002/jps.20322
\item Rodgers/Rowland (2006), DOI: 10.1002/jps.20502
}
}
