% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{setopt_compphysiol}
\alias{setopt_compphysiol}
\title{Set package options}
\usage{
setopt_compphysiol(...)
}
\arguments{
\item{...}{Named arguments to update package options}
}
\description{
Set package options
}
