% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{sMD_PBPK_xCMT_lumped}
\alias{sMD_PBPK_xCMT_lumped}
\title{Lumped PBPK model based on sMD_PBPK_12CMT_wellstirred}
\usage{
sMD_PBPK_xCMT_lumped(partitioning, autonormalize = TRUE)
}
\arguments{
\item{partitioning}{A named list defining the lumping scheme}

\item{autonormalize}{Whether to automatically normalize the lumped volumes}
}
\value{
A CompartmentModel object
}
\description{
Lumped PBPK model based on sMD_PBPK_12CMT_wellstirred
}
