% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartments.R
\name{states}
\alias{states}
\title{States method for \code{Compartments} class
Returns the states of compartments in a \code{Compartments} object.}
\usage{
states(comp)
}
\arguments{
\item{comp}{A \code{Compartments} object}
}
\value{
A character vector of compartment states
}
\description{
States method for \code{Compartments} class
Returns the states of compartments in a \code{Compartments} object.
}
