% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSL.R
\name{[.compphysiol_dsl}
\alias{[.compphysiol_dsl}
\title{Subsetting expressions of the form \code{a[molec,cmt]} and \code{c[molec,cmt]}}
\usage{
\method{[}{compphysiol_dsl}(x, molec, cmt)
}
\arguments{
\item{x}{A \code{compphysiol_dsl} object}

\item{molec}{A molecule name (symbol)}

\item{cmt}{A compartment name (symbol)}
}
\value{
The corresponding value \code{a[molec,cmt]} or \code{c[molec,cmt]} retrieved
from the evaluation environment.
}
\description{
Subsetting expressions of the form \code{a[molec,cmt]} and \code{c[molec,cmt]}
}
