% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drugs.R
\name{[.Drugs}
\alias{[.Drugs}
\title{Subset method for Drugs objects}
\usage{
\method{[}{Drugs}(x, i)
}
\arguments{
\item{x}{An object of class \code{Drugs}.}

\item{i}{Index(es) or name(s) of drug(s) to select}
}
\value{
A subsetted \code{Drugs} object.
}
\description{
This method allows you to subset a \code{Drugs} object using standard indexing.
}
