% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiments.R
\name{[.Experiments}
\alias{[.Experiments}
\title{Subset a collection of experiments}
\usage{
\method{[}{Experiments}(x, i, ...)
}
\arguments{
\item{x}{An \code{Experiments} object.}

\item{i}{Numeric, logical, or character indices.}

\item{...}{Unused.}
}
\value{
An \code{Experiments} object.
}
\description{
Missing or unknown indices that would introduce missing elements are rejected.
}
