% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flows.R
\name{[.Flows}
\alias{[.Flows}
\title{Subset a \code{Flows} object}
\usage{
\method{[}{Flows}(x, i, ...)
}
\arguments{
\item{x}{A \code{Flows} object}

\item{i}{Row indices to subset}

\item{...}{Additional arguments (not used)}
}
\value{
A subsetted \code{Flows} object
}
\description{
Subset a \code{Flows} object
}
