% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecules.R
\name{[.Molecules}
\alias{[.Molecules}
\title{Subset a \code{Molecules} object}
\usage{
\method{[}{Molecules}(x, i, ...)
}
\arguments{
\item{x}{A \code{Molecules} object}

\item{i}{Index or name(s) of molecules to subset}

\item{...}{Additional arguments (not used)}
}
\value{
A subsetted \code{Molecules} object
}
\description{
Subset a \code{Molecules} object
}
