% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{[.ParameterEstimates}
\alias{[.ParameterEstimates}
\title{Subset parameter-estimation specifications}
\usage{
\method{[}{ParameterEstimates}(x, i, ...)
}
\arguments{
\item{x}{A \code{ParameterEstimates} collection.}

\item{i}{Indices or names of estimates to retain.}

\item{...}{Unused.}
}
\value{
A \code{ParameterEstimates} collection.
}
\description{
Subset parameter-estimation specifications
}
