% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSets.R
\name{[.ParameterSets}
\alias{[.ParameterSets}
\title{Subset realized parameter sets}
\usage{
\method{[}{ParameterSets}(x, i, ...)
}
\arguments{
\item{x}{A \code{ParameterSets} object.}

\item{i}{Numeric, logical, or character indices.}

\item{...}{Unused.}
}
\value{
A \code{ParameterSets} object.
}
\description{
Subset realized parameter sets
}
