% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{[.ParameterSpecs}
\alias{[.ParameterSpecs}
\title{Subset parameter-estimation specifications}
\usage{
\method{[}{ParameterSpecs}(x, i, ...)
}
\arguments{
\item{x}{A \code{ParameterSpecs} collection.}

\item{i}{Indices or names of specifications to retain.}

\item{...}{Unused.}
}
\value{
A \code{ParameterSpecs} collection.
}
\description{
Subset parameter-estimation specifications
}
