% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalModel.R
\name{[.StatisticalModel}
\alias{[.StatisticalModel}
\title{Subset a statistical model}
\usage{
\method{[}{StatisticalModel}(x, i, ...)
}
\arguments{
\item{x}{A \code{StatisticalModel} object.}

\item{i}{Indices or names of random quantities to retain.}

\item{...}{Unused.}
}
\value{
A \code{StatisticalModel} object.
}
\description{
Subset a statistical model
}
