% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transports.R
\name{[.Transports}
\alias{[.Transports}
\title{Subset a \code{Transports} object}
\usage{
\method{[}{Transports}(x, i, ...)
}
\arguments{
\item{x}{A \code{Transports} object}

\item{i}{Row indices to subset}

\item{...}{Additional arguments (not used)}
}
\value{
A subsetted \code{Transports} object
}
\description{
Subset a \code{Transports} object
}
