% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drugs.R
\name{[[.Drugs}
\alias{[[.Drugs}
\title{Extract method for Drugs objects.}
\usage{
\method{[[}{Drugs}(x, i)
}
\arguments{
\item{x}{An object of class \code{Drugs}.}

\item{i}{The name of the drug to extract.}
}
\value{
A \code{Drug} object containing the parameters for the specified drug.
}
\description{
Extract method for Drugs objects.
}
