% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flows.R
\name{[[.Flows}
\alias{[[.Flows}
\title{Accessing an element in a \code{Flows} object}
\usage{
\method{[[}{Flows}(x, i)
}
\arguments{
\item{x}{A \code{Flows} object}

\item{i}{Row index to access}
}
\value{
A \code{Flow} object, which is a list with entries \code{from}, \code{to}, \code{rate}, \code{const}, and \code{type}.
}
\description{
Accessing an element in a \code{Flows} object
}
