% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transports.R
\name{[[.Transports}
\alias{[[.Transports}
\title{Access an element in a \code{Transports} object}
\usage{
\method{[[}{Transports}(x, i)
}
\arguments{
\item{x}{A \code{Transports} object}

\item{i}{Row index to access}
}
\value{
A \code{Transport} object, which is a list with entries \code{from}, \code{to}, \code{molec}, \code{rate}, \code{const}, and \code{type}.
}
\description{
Access an element in a \code{Transports} object
}
