% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{[<-.Parameters}
\alias{[<-.Parameters}
\title{Replace parts of a \code{Parameters} object}
\usage{
\method{[}{Parameters}(x, i) <- value
}
\arguments{
\item{x}{A \code{Parameters} object}

\item{i}{Indices or names of the parameters to replace}

\item{value}{Another \code{Parameters} object with the new values to insert}
}
\value{
The modified \code{Parameters} object
}
\description{
Replace parts of a \code{Parameters} object
}
