% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression-handling.R
\name{substitute_expr}
\alias{substitute_expr}
\title{Substitute states and parameters in an expression}
\usage{
substitute_expr(
  expr,
  stateNames,
  name2idx,
  paramValues = list(),
  freeParamsEnv = NULL,
  obsFunc = FALSE
)
}
\arguments{
\item{expr}{character string or R expression}

\item{stateNames}{character vector of state names}

\item{name2idx}{named integer mapping from stateNames -> index}

\item{paramValues}{named list of parameter values (scalars may be inlined)}

\item{obsFunc}{logical, TRUE if in observable (matrix indexing)}

\item{freeParams}{character vector (will be updated by reference)}
}
\value{
An R expression with substitutions applied
}
\description{
Substitute states and parameters in an expression
}
