% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{summary.Physiology}
\alias{summary.Physiology}
\title{Summarize a \code{Physiology} object as a string.}
\usage{
\method{summary}{Physiology}(object, ...)
}
\arguments{
\item{object}{A \code{Physiology} object}

\item{...}{Additional arguments (not used)}
}
\value{
A summary string
}
\description{
Summarize a \code{Physiology} object as a string.
}
