% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{to_param_list}
\alias{to_param_list}
\title{Export parameters as a named list}
\usage{
to_param_list(phys)
}
\arguments{
\item{phys}{A \code{Physiology} object}
}
\value{
A named list of parameters and metadata
}
\description{
This function converts the parameters in the \code{param_table} of a \code{Physiology} object into a named list,
where the names are either the parameter name (for scalars) or \code{parameter[context]} (for tissue parameters).
}
