% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{wire}
\alias{wire}
\title{Wire wildcard molecules and / or compartments into a compartment model}
\usage{
wire(model, what = c("molec", "cmt"))
}
\arguments{
\item{model}{A \code{CompartmentModel} object}

\item{what}{What to wire: \code{"molec"} for molecules, \code{"cmt"} for compartments, or both (the default)}
}
\description{
The \code{wire()} function allows users to replace wildcard molecule and compartment names
in transport and reaction definitions with actual names from the model.
A fully wired model is safe from any potentially unintended side effects when extending the model.
\code{wire()} is also called before ODE generation to replace any remaining wildcards with the full set of molecule or compartment names.
}
\details{
If molecules / compartments have been defined in the model, then any transport or reaction that references the wildcard name
"\if{html}{\out{<all molec>}}" / "\if{html}{\out{<all cmt>}}" will be updated to reference all defined molecules / compartments, respectively
(note that this may increase the number of molecules, transports and reactions in the model).
If no compartments or molecules have been defined, then a single dummy name ("molec" or "cmt") will be used
to replace the wildcard, with a message printed to the console to inform the user.
In addition, rate expressions and observables that include statements of the form \verb{[molec]} or \verb{[cmt]} will be
expanded to include the implicit molecule and compartment names (i.e., into the form \verb{[molec,cmt]}).
}
\examples{
# wire "drug" into transport rate expressions, replacing the <all molec> wildcard
M <- compartment_model() |>
  add_compartment(name = c("cen", "per")) |>
  add_molecule(name = "drug") |>
  add_transport(from = "cen", to = "per", const = "kcp") |> 
  add_transport(from = "per", to = "cen", const = "kpc") |> 
  wire(what = "molec")

# wire "cyt" and "nuc" into molecule definitions and reaction rates,
# replacing the <all cmt> wildcard
M <- compartment_model() |>
  add_compartment(name = c("cyt", "nuc")) |>
  add_molecule(name = c("A","B")) |>
  add_reaction(input = "A", output = "B", const = "kAB") |>
  wire(what = "cmt")
}
