% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{add_compartment}
\alias{add_compartment}
\title{Add one or several compartments to a \code{CompartmentModel} object.}
\usage{
add_compartment(
  model,
  name = character(0),
  initial = 0,
  unit = NULL,
  state = paste0("A", name, recycle0 = TRUE),
  comp
)
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}

\item{name}{Compartment name(s), character scalar or vector}

\item{unit}{Unit for the compartment volume(s) (optional, default: \code{NULL}). If specified,
it can be a single unit (applied to all volumes) or a vector of units corresponding to each volume.
The \code{unit} argument is ignored (with a warning) if the \code{volume} argument is not numeric.}

\item{comp}{A \code{Compartments} object. Constructed from the other inputs if not provided.}
}
\value{
The modified \code{CompartmentModel} object.
}
\description{
Compartments represent model compartments with names and fixed or parametrized volumes.
}
\examples{
# Several compartments added in a single call
compartment_model() |>
    add_compartment(c("A","B"), unit = "mg")

# Different ways of adding compartments
M1 <- compartment_model() |>
    add_compartment("A", initial = 10 [mg])
M2 <- compartment_model() |>
    add_compartment("A", initial = 10, unit = "mg")
identical(M1, M2)
}
