% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{add_compartment}
\alias{add_compartment}
\title{Add one or several compartments to a \code{CompartmentModel} object.}
\usage{
add_compartment(model, name, initial = 0, comp = compartments(name, initial))
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}

\item{name}{Name of the compartment(s)}

\item{initial}{Initial amount(s) (default 0)}

\item{comp}{A \code{Compartments} object. Constructed from the other inputs if not provided.}
}
\description{
Add one or several compartments to a \code{CompartmentModel} object.
}
