% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartments.R
\name{add_compartment}
\alias{add_compartment}
\title{Add one or several compartments to a \code{CompartmentModel} object.}
\usage{
add_compartment(
  model,
  name = character(0),
  volume = "V{name}",
  unit = NULL,
  comp
)
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}

\item{name}{Compartment name(s), character scalar or vector}

\item{volume}{Volume(s) (numeric or character) for each compartment.
The default \verb{V\{name\}} uses parametrized volumes, with \code{{name}} replaced by the compartment name(s).
Parametrized volumes have to be provided via the \code{parameters} argument of \code{compartment_model()}, e.g., \code{parameters(Vadi = 5 [L])}.
Fixed numeric volumes can also be provided, e.g., \code{volume = 5 [L]} or \verb{volume = 5, unit = "L"}.
Can be a single value (applied to all compartments) or a vector of values corresponding to each compartment.}

\item{unit}{Unit for the compartment volume(s) (optional, default: \code{NULL}). If specified,
it can be a single unit (applied to all volumes) or a vector of units corresponding to each volume.
The \code{unit} argument is ignored (with a warning) if the \code{volume} argument is not numeric.}

\item{comp}{A \code{Compartments} object. Constructed from the other inputs if not provided.}
}
\value{
The modified \code{CompartmentModel} object.
}
\description{
Compartments represent model compartments with names and fixed or parametrized volumes.
}
\examples{
# Several compartments added in a single call
compartment_model() |>
    add_compartment(c("A","B"))

# Different ways of adding compartments
M1 <- compartment_model() |>
    add_compartment("A", volume = 10 [L])
M2 <- compartment_model() |>
    add_compartment("A", volume = 10, unit = "L")
identical(M1, M2)
}
