% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{add_compartment}
\alias{add_compartment}
\title{Add one or several compartments to a \code{CompartmentModel} object.}
\usage{
add_compartment(
  model,
  name,
  initial = 0,
  unit = NULL,
  state = paste0("A", name, recycle0 = TRUE),
  comp
)
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}

\item{name}{Compartment name(s), character scalar or vector}

\item{initial}{Initial amounts for the compartments (numeric scalar or vector, default = \code{0})}

\item{unit}{Units for the initial amounts (character scalar or vector, default = \code{NULL}).
If provided, must be either a single unit applied to all compartments or a vector of units matching the length of \code{name}.}

\item{state}{State names for the compartments (character scalar or vector, default = \code{"A" + name}).}

\item{comp}{A \code{Compartments} object. Constructed from the other inputs if not provided.}
}
\value{
The modified \code{CompartmentModel} object.
}
\description{
Add one or several compartments to a \code{CompartmentModel} object.
}
