% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{add_dosing}
\alias{add_dosing}
\title{Add one or several dosing events (bolus or infusion).}
\usage{
add_dosing(
  model,
  target,
  time,
  amount = NULL,
  ...,
  rate = NULL,
  duration = NULL,
  dose = dosing(target = target, time = time, amount = amount, rate = rate, duration =
    duration)
)
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}

\item{target}{Name of the target compartment(s) for the dose(s)}

\item{time}{Time of the dosing event(s)}

\item{amount}{Amount(s) to be dosed (for bolus or infusion)}

\item{...}{Unused, enforces \code{rate} / \code{duration} / \code{dose} to be specified as named arguments only, not positional}

\item{rate}{Infusion rate (for infusion)}

\item{duration}{Infusion duration (for infusion)}

\item{dose}{A \code{Dosing} object. Constructed from the other inputs if not provided.}
}
\value{
The modified \code{CompartmentModel} object.
}
\description{
This function allows the user to specify dosing events for a compartment model.
Dosing events can be either bolus (instantaneous) or infusion (continuous over time).
}
\details{
The function handles the necessary modifications to the model structure for infusion dosing,
such as adding infusion bag and rate compartments, and creating the appropriate events for
starting and stopping the infusion.

TODO: Currently there is an infinite loop caused by the pipeline of \code{add_dosing} calling \code{add_dosing}
when processing infusion dosing. This is a temporary issue that will be resolved in the next iteration
of the code refactor, where the internal handling of infusion dosing will be separated from the user-facing
\code{add_dosing} function.
}
\examples{
model <- compartment_model() |>
    add_dosing(target = "ven", time = 0, amount = 100, duration = 5)
}
