% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{add_dosing}
\alias{add_dosing}
\title{Add one or several dosing events (bolus or infusion).}
\usage{
add_dosing(
  model,
  target,
  time,
  amount = NULL,
  time_unit = NULL,
  amount_unit = NULL,
  ...,
  rate = NULL,
  duration = NULL,
  dose
)
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}

\item{target}{Target compartment name (character scalar)}

\item{time}{Dosing time(s) (numeric scalar or vector)}

\item{amount}{Amount(s) of dose (non-negative numeric scalar or vector, optional)}

\item{time_unit}{Optional unit for time (character scalar, e.g., "h", or \code{NULL}, the default)}

\item{amount_unit}{Optional unit for amount (character scalar, e.g., "mg", or \code{NULL}, the default)}

\item{...}{Unused, enforces \code{rate} and \code{duration} to be specified as named arguments only, not positional}

\item{rate}{Infusion rate (non-negative numeric scalar or vector, optional)}

\item{duration}{Infusion duration (positive numeric scalar or vector, optional)}

\item{dose}{A \code{Dosing} object. Constructed from the other inputs if not provided.}
}
\value{
The modified \code{CompartmentModel} object.
}
\description{
This function allows the user to specify dosing events for a compartment model.
Dosing events can be either bolus (instantaneous) or infusion (continuous over time).
}
\details{
The function handles the necessary modifications to the model structure for infusion dosing,
such as adding infusion bag and rate compartments, and creating the appropriate events for
starting and stopping the infusion.
}
\examples{
model <- compartment_model() |>
    add_dosing(target = "ven", time = 0, amount = 100, duration = 5)
}
