% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{add_dosing}
\alias{add_dosing}
\title{Add one or several dosing events (bolus or infusion).}
\usage{
add_dosing(
  model,
  target,
  time,
  amount = NULL,
  rate = NULL,
  duration = NULL,
  dose = DosingList(target = target, time = time, amount = amount, rate = rate, duration
    = duration)
)
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}

\item{target}{Name of the target compartment(s) for the dose(s)}

\item{time}{Time of the dosing event(s)}

\item{amount}{Amount(s) to be dosed (for bolus) or infusion rate (for infusion)}

\item{rate}{Infusion rate (for infusion)}

\item{duration}{Infusion duration (for infusion)}

\item{dose}{A Dosing object or a list of Dosing objects. Constructed from the other inputs if not provided.}
}
\value{
The modified \code{CompartmentModel} object.
}
\description{
Add one or several dosing events (bolus or infusion).
}
\examples{
model <- compartment_model() |> 
    add_dosing(target = "ven", time = 0, amount = 100, duration = 5)
}
