% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{add_equation}
\alias{add_equation}
\title{Add one or several equations to a \code{CompartmentModel} object.}
\usage{
add_equation(model, ..., eq = NULL)
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}

\item{...}{Name-expression pairs for equations, e.g. \code{co = Qadi+Qbon+Qhea+Qkid+Qliv+Qmus+Qski}}

\item{eq}{An \code{Equations} object.}
}
\value{
The modified \code{CompartmentModel} object.
}
\description{
Add one or several equations to a \code{CompartmentModel} object.
}
\examples{
## Interactive path with name-expression pairs
compartment_model() |>
    add_equation(co = Qadi+Qbon+Qhea+Qkid+Qliv+Qmus+Qski)
## Programmatic path with Equations object
eq <- equations(name = "co", expr = "Qadi+Qbon+Qhea+Qkid+Qliv+Qmus+Qski")
compartment_model() |>
    add_equation(eq = eq)
}
