% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{add_equation}
\alias{add_equation}
\title{Add one or several equations to a \code{CompartmentModel} object.}
\usage{
add_equation(model, eq)
}
\arguments{
\item{eq}{An Equation object or list of Equation objects. Constructed from the other inputs if not provided.}
}
\value{
The modified \code{CompartmentModel} object.
}
\description{
Add one or several equations to a \code{CompartmentModel} object.
}
\examples{
model <- compartment_model() |>
    add_equation("co = Qadi+Qbon+Qgut+Qhea+Qkid+Qliv+Qmus+Qski+Qspl")
}
