% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{add_flow}
\alias{add_flow}
\title{Add one or several flows to a \code{CompartmentModel} object.}
\usage{
add_flow(
  model,
  from,
  to,
  rate = NULL,
  const = NULL,
  flow = flows(from, to, rate, const)
)
}
\arguments{
\item{from}{Source compartment}

\item{to}{Target compartment}

\item{rate}{Optional flow rate expression as character}

\item{const}{Optional rate constant name (for linear flows)}

\item{flow}{A \code{Flows} object. Constructed from the other inputs if not provided.}
}
\value{
The modified \code{CompartmentModel} object.
}
\description{
Add one or several flows to a \code{CompartmentModel} object.
}
\examples{
model <- compartment_model() |>
    add_flow(from = "A", to = "B", const = "k1")
}
