% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecules.R
\name{add_molecule}
\alias{add_molecule}
\title{Add one or several molecules to a \code{CompartmentModel} object.}
\usage{
add_molecule(
  model,
  name = character(0),
  cmt = NULL,
  initial = 0,
  unit = NULL,
  type = c("concentration", "amount"),
  molec = NULL
)
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}

\item{name}{Molecule name(s), character scalar or vector}

\item{cmt}{Compartment name(s), character scalar or vector}

\item{initial}{Initial concentration(s)/amount(s) of the molecule (numeric, possibly as \code{values[unit]}). The dimension of the
\code{initial} argument needs to be compatible with the \code{molec} and \code{cmt} arguments (to be clarified).}

\item{unit}{Unit for the initial conditions (character scalar or \code{NULL}, the default), to be used if units are not specified via \code{initial}.}

\item{type}{Type of initial condition, either \code{"concentration"} (the default) or \code{"amount"}.}

\item{molec}{A \code{Molecules} object. Constructed from the other inputs if not provided.}
}
\value{
The modified \code{CompartmentModel} object.
}
\description{
Add one or several molecules to a \code{CompartmentModel} object.
}
\seealso{
\code{\link[=molecules]{molecules()}}
}
