% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{add_observable}
\alias{add_observable}
\title{Add an observable to a \code{CompartmentModel} object.}
\usage{
add_observable(model, ..., obs = NULL)
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}

\item{...}{Name-expression pairs for observables, e.g. \code{Cblo = blo / Vblo}}

\item{obs}{An \code{Observables} object.}
}
\value{
The modified \code{CompartmentModel} object.
}
\description{
Add an observable to a \code{CompartmentModel} object.
}
\examples{
## Interactive path with name-expression pairs
compartment_model() |>
    add_compartment("blo") |>
    add_observable(Cblo = blo/Vblo)
## Programmatic path with Observables object
obs <- observables(name = "Cblo", expr = "blo/Vblo")
compartment_model() |>
    add_compartment("blo") |>
    add_observable(obs = obs)
}
