% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{add_observable}
\alias{add_observable}
\title{Add an observable to a \code{CompartmentModel} object.}
\usage{
add_observable(model, name, expr, obs = ObservableList(name, expr))
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}

\item{name}{Name of the observable}

\item{expr}{Expression (character or function)}

\item{obs}{An Observable object or list of Observable objects. Constructed from the other inputs if not provided.}
}
\value{
The modified \code{CompartmentModel} object.
}
\description{
Add an observable to a \code{CompartmentModel} object.
}
\examples{
model <- compartment_model() |>
    add_compartment("blo") |>
    add_observable(name = "Cblo", expr = "blo/Vblo")
}
