% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{add_parameter}
\alias{add_parameter}
\title{Add one or several parameters to a \code{CompartmentModel} object.
Parameters can be added interactively as name-value pairs (potentially with units), or programmatically as a \code{Parameters} object.}
\usage{
add_parameter(model, ..., name = NULL, value = NULL, unit = NULL, param)
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}

\item{...}{Parameter values as name-value pairs, where values can optionally have units (e.g., \code{A = 2 [m]}).}

\item{name}{Optional parameter names (if not using named arguments).}

\item{value}{Optional parameter values (if not using named arguments).}

\item{unit}{Optional parameter units (if not using named arguments).}

\item{param}{A \code{Parameters} object. Constructed from the other inputs if not provided.}
}
\value{
The modified \code{CompartmentModel} object.
}
\description{
Add one or several parameters to a \code{CompartmentModel} object.
Parameters can be added interactively as name-value pairs (potentially with units), or programmatically as a \code{Parameters} object.
}
