% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transports.R
\name{add_transport}
\alias{add_transport}
\title{Add one or several transports to a \code{CompartmentModel} object.}
\usage{
add_transport(model, from, to, ..., rate = NULL, const = NULL, trans)
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}

\item{from}{The name of the source compartment(s) (character vector or NULL for source compartments)}

\item{to}{The name of the destination compartment(s) (character vector or NULL for sink compartments)}

\item{...}{Unused, enforces \code{rate} and \code{const} to be specified as named arguments only, not positional}

\item{rate}{The transport rate (character or call, optional, mutually exclusive with \code{const})}

\item{const}{Rate constant for first-order transports (character or call, optional, mutually exclusive with \code{rate})}

\item{trans}{A \code{Transports} object. Constructed from the other inputs if not provided.}
}
\value{
The modified \code{CompartmentModel} object.
}
\description{
Transports represent mass/amount-preserving transports of molecules between two compartments in a compartment model.
}
\details{
Transports can be linear (first-order) or nonlinear, depending on whether the transport rate is defined by a constant
(argument \code{const}) or a general expression (argument \code{rate}).
The special syntax \code{{from}}, \code{{to}} and \code{{molec}} can be used in \code{rate} or \code{const} to automatically substitute
the source and destination compartment names and molecule names, respectively, e.g., \code{"k_{from}_{to}"} will become
\code{k_ven_bon} for a transport from the venous blood (ven) to bone (bon). This allows for vector definition of multiple
transports with different compartments and/or molecules involved.
}
\examples{
# Add a linear transport
compartment_model() |>
    add_transport(from = "A", to = "B", const = "k1")
# Add a nonlinear transport
compartment_model() |>
    add_transport(from = "A", to = "B", rate = "k2 * c[A] / (c[A] + Km)")
# Vectorized calls using naming abbreviation
compartment_model() |>
    add_transport("A", c("B", "C"), const = "k{to}")
compartment_model() |>
    add_transport(c("A", "B"), c("B", "C"), const = "k_{from}_{to}")
# Transports with specific molecules
compartment_model() |>
    add_transport(from = "cen", to = "per", molec = c("DrugA","DrugB"), const = "k_{molec}")
}
