% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartments.R
\name{c.Compartments}
\alias{c.Compartments}
\title{Combine multiple \code{Compartments} objects into one}
\usage{
\method{c}{Compartments}(...)
}
\arguments{
\item{...}{Multiple \code{Compartments} objects to combine}
}
\value{
A combined \code{Compartments} object
}
\description{
Combine multiple \code{Compartments} objects into one
}
