% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartments.R
\name{Compartment}
\alias{Compartment}
\title{Compartment Class}
\description{
Represents a model compartment with a name and an initial amount.
}
\examples{
cpt <- Compartment$new("Heart", 10)
cpt$name
cpt$initial

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Character string giving the name of the compartment.}

\item{\code{initial}}{Numeric value, the initial amount in the compartment.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Compartment-new}{\code{Compartment$new()}}
\item \href{#method-Compartment-print}{\code{Compartment$print()}}
\item \href{#method-Compartment-clone}{\code{Compartment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compartment-new"></a>}}
\if{latex}{\out{\hypertarget{method-Compartment-new}{}}}
\subsection{Method \code{new()}}{
Create a new compartment object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$new(name, initial = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the compartment (character)}

\item{\code{initial}}{Initial amount (numeric, default = 0)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Compartment} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compartment-print"></a>}}
\if{latex}{\out{\hypertarget{method-Compartment-print}{}}}
\subsection{Method \code{print()}}{
Print a compartment object to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments (not used)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{Compartment} object (invisible)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compartment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Compartment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
