% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{compartment_model}
\alias{compartment_model}
\title{Create a new CompartmentModel object.}
\usage{
compartment_model()
}
\value{
A new \code{CompartmentModel} object.
}
\description{
Create a new CompartmentModel object.
}
