% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{CompartmentModel}
\alias{CompartmentModel}
\title{CompartmentModel class}
\description{
Represents a compartmental model.
}
\details{
\code{CompartmentModel} is an R6 class that encapsulates the structure of a compartmental model,
including its compartments, reactions, observables, and dosing events.
}
\examples{

## ------------------------------------------------
## Method `CompartmentModel$print`
## ------------------------------------------------

M <- multiCompModel()
print(M)

## ------------------------------------------------
## Method `CompartmentModel$toAnalytical`
## ------------------------------------------------

M <- multiCompModel(ncomp = 2, type = "micro")
sol <- M$toAnalytical(paramValues = list(k10 = 0.05))  # fix one param

# Evaluate ODE state at t = 5 with free params
sol$statefun(5, params = list(k12 = 0.2, k21 = 0.1))
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{compartments}}{A list of \code{Compartment} objects}

\item{\code{reactions}}{A list of \code{Reaction} objects}

\item{\code{equations}}{A list of expressions representing aggregate parameters as functions of model parameters}

\item{\code{observables}}{A list of \code{Observable} objects}

\item{\code{doses}}{List of dosing objects. Includes bolus doses into model
compartments, as well as bolus-to-bag doses used to implement infusions.}

\item{\code{infusionEvents}}{Data frame of infusion rate adjustment events
(start and stop of zero-order input). Generated automatically when
infusions are defined.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CompartmentModel-new}{\code{CompartmentModel$new()}}
\item \href{#method-CompartmentModel-print}{\code{CompartmentModel$print()}}
\item \href{#method-CompartmentModel-getInitialState}{\code{CompartmentModel$getInitialState()}}
\item \href{#method-CompartmentModel-getStateNames}{\code{CompartmentModel$getStateNames()}}
\item \href{#method-CompartmentModel-addCompartment}{\code{CompartmentModel$addCompartment()}}
\item \href{#method-CompartmentModel-addReaction}{\code{CompartmentModel$addReaction()}}
\item \href{#method-CompartmentModel-addObservableOld}{\code{CompartmentModel$addObservableOld()}}
\item \href{#method-CompartmentModel-addObservable}{\code{CompartmentModel$addObservable()}}
\item \href{#method-CompartmentModel-addEquation}{\code{CompartmentModel$addEquation()}}
\item \href{#method-CompartmentModel-addDosing}{\code{CompartmentModel$addDosing()}}
\item \href{#method-CompartmentModel-isLinear}{\code{CompartmentModel$isLinear()}}
\item \href{#method-CompartmentModel-toAnalytical}{\code{CompartmentModel$toAnalytical()}}
\item \href{#method-CompartmentModel-toODE}{\code{CompartmentModel$toODE()}}
\item \href{#method-CompartmentModel-clone}{\code{CompartmentModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new \code{CompartmentModel} object.
Print method for CompartmentModel class

Pretty-prints a \code{CompartmentModel} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-print"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{CompartmentModel} object (invisibly).
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{M <- multiCompModel()
print(M)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-getInitialState"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-getInitialState}{}}}
\subsection{Method \code{getInitialState()}}{
Get initial states for simulation as a named vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$getInitialState(named = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{named}}{A boolean, should the initial states be named?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-getStateNames"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-getStateNames}{}}}
\subsection{Method \code{getStateNames()}}{
Get compartment names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$getStateNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-addCompartment"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-addCompartment}{}}}
\subsection{Method \code{addCompartment()}}{
Add one or several compartments to the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$addCompartment(
  name,
  initial = 0,
  comp = CompartmentList(name, initial)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the compartment(s)}

\item{\code{initial}}{Initial amount(s) (default 0)}

\item{\code{comp}}{A Compartment object or list of Compartment objects. Constructed from the other inputs if not provided.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-addReaction"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-addReaction}{}}}
\subsection{Method \code{addReaction()}}{
Add one or several reactions to the model.
Currently, only the \code{reaction} argument can be vectorized.
The other inputs can only be used to construct a scalar reaction.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$addReaction(
  from,
  to,
  rate = NULL,
  const = NULL,
  reaction = Reaction$new(from, to, rate, const)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from}}{Source compartment}

\item{\code{to}}{Target compartment}

\item{\code{rate}}{Rate expression as character}

\item{\code{const}}{Optional rate constant name (for linear reactions)}

\item{\code{reaction}}{A Reaction object or list of Reaction objects. Constructed from the other inputs if not provided.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-addObservableOld"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-addObservableOld}{}}}
\subsection{Method \code{addObservableOld()}}{
Add an observable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$addObservableOld(name, expr, obs = ObservableList(name, expr))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the observable}

\item{\code{expr}}{Expression (character or function)}

\item{\code{obs}}{An Observable object or list of Observable objects. Constructed from the other inputs if not provided.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-addObservable"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-addObservable}{}}}
\subsection{Method \code{addObservable()}}{
Add one or several observables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$addObservable(..., obs = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Name and expression pairs for observables, e.g. \code{addObservable(C1Conc = C1 / V1)}}

\item{\code{obs}}{An Observable object or list of Observable objects. Constructed from the other inputs if not provided.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-addEquation"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-addEquation}{}}}
\subsection{Method \code{addEquation()}}{
Add one or several equations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$addEquation(eq)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{eq}}{An Equation object or list of Equation objects. Constructed from the other inputs if not provided.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-addDosing"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-addDosing}{}}}
\subsection{Method \code{addDosing()}}{
Add one or several dosing events (bolus or infusion).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$addDosing(
  target,
  time,
  amount = NULL,
  rate = NULL,
  duration = NULL,
  dose = DosingList(target = target, time = time, amount = amount, rate = rate, duration
    = duration)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target}}{Name of the target compartment(s) for the dose(s)}

\item{\code{time}}{Time of the dosing event(s)}

\item{\code{amount}}{Amount(s) to be dosed (for bolus) or infusion rate (for infusion)}

\item{\code{rate}}{Infusion rate (for infusion)}

\item{\code{duration}}{Infusion duration (for infusion)}

\item{\code{dose}}{A Dosing object or a list of Dosing objects. Constructed from the other inputs if not provided.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-isLinear"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-isLinear}{}}}
\subsection{Method \code{isLinear()}}{
Linearity check for CompartmentModel object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$isLinear()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if all reactions are linear, \code{FALSE} otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-toAnalytical"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-toAnalytical}{}}}
\subsection{Method \code{toAnalytical()}}{
Generate analytical solution function from a CompartmentModel object with
linear reactions, up to 3 compartments and a single bolus dose at time 0.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$toAnalytical(paramValues = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{paramValues}}{Named list of parameter values to inline in ODEs.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A length 2 list named \code{state} (a function) and \code{observable}
(a list of functions, possibly empty). \code{state(t,param)} calculates the ODE
solution at time \code{t} for free parameters \code{param}, while \code{observable[[i]](t,param)}
calculates the \code{i}-th observable defined in the CompartmentModel.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{M <- multiCompModel(ncomp = 2, type = "micro")
sol <- M$toAnalytical(paramValues = list(k10 = 0.05))  # fix one param

# Evaluate ODE state at t = 5 with free params
sol$statefun(5, params = list(k12 = 0.2, k21 = 0.1))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-toODE"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-toODE}{}}}
\subsection{Method \code{toODE()}}{
Generate ODE function, initial values, observables, and free parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$toODE(paramValues = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{paramValues}}{Named list of parameter values to inline in ODEs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompartmentModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CompartmentModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompartmentModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
