% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartments.R
\name{compartments}
\alias{compartments}
\title{\code{Compartments} Class}
\usage{
compartments(name, initial = 0)
}
\arguments{
\item{name}{Compartment name(s), character scalar or vector}

\item{initial}{Initial amounts for the compartments (numeric scalar or vector, default = 0)}
}
\value{
A \code{Compartments} object
}
\description{
Represents one or several model compartments with names and initial amounts.
}
\examples{
compartments(c("adi","bon"), 10)
}
