% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compphysiol-package.R
\docType{package}
\name{compphysiol}
\alias{compphysiol-package}
\alias{compphysiol}
\title{compphysiol: A Package for Computational Physiology}
\description{
The \code{compphysiol} package provides tools for building and simulating compartment
models in R. It includes classes for defining compartments, reactions, observables,
dosing events, and physiological parameters.
}
\details{
Key classes include:
\itemize{
\item \code{CompartmentModel}: Encapsulates the structure of a compartmental model,
including compartments, reactions, observables, and dosing events.
\item \code{Compartments}: Represents model compartments with names and initial amounts.
\item \code{Flows}: Defines flows between compartments with specified rates.
\item \code{Observables}: Represents measurable outputs derived from compartment states.
\item \code{Equations}: Represents mathematical relationships between model parameters and/or states.
\item \code{Dosing}: Represents dosing events, including bolus and infusion doses.
\item \code{Physiology}: Manages physiological parameters for the model.
}
}
\author{
\strong{Maintainer}: Niklas Hartung \email{niklas.hartung@uni-potsdam.de}

}
\keyword{internal}
