% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dosing.R
\name{Dosing}
\alias{Dosing}
\title{Dosing Class}
\description{
Represents dosing into a CompartmentModel
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Dosing-new}{\code{Dosing$new()}}
\item \href{#method-Dosing-isBolus}{\code{Dosing$isBolus()}}
\item \href{#method-Dosing-isInfusion}{\code{Dosing$isInfusion()}}
\item \href{#method-Dosing-clone}{\code{Dosing$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dosing-new"></a>}}
\if{latex}{\out{\hypertarget{method-Dosing-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dosing$new(target, amount = NULL, time, rate = NULL, duration = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dosing-isBolus"></a>}}
\if{latex}{\out{\hypertarget{method-Dosing-isBolus}{}}}
\subsection{Method \code{isBolus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dosing$isBolus()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dosing-isInfusion"></a>}}
\if{latex}{\out{\hypertarget{method-Dosing-isInfusion}{}}}
\subsection{Method \code{isInfusion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dosing$isInfusion()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dosing-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Dosing-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dosing$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
