% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dosing.R
\name{DosingList}
\alias{DosingList}
\title{Construct one or more Dosing objects}
\usage{
DosingList(target, time, amount = NULL, rate = NULL, duration = NULL)
}
\description{
If \code{time} (and optionally \code{amount}, etc.) are vectors, returns
a list of \code{Dosing} objects. Otherwise, returns a single Dosing.
}
\examples{
DosingList("Central", time = c(0, 24, 48), amount = 100)
DosingList("Central", time = 0, amount = 50)

}
