% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equations.R
\name{equations}
\alias{equations}
\title{Create an \code{Equations} object}
\usage{
equations(name, expr)
}
\arguments{
\item{name}{Name of the equation(s), character scalar or vector}

\item{expr}{Expression (as an R call)}
}
\value{
An \code{Equations} object
}
\description{
Create an \code{Equations} object
}
\examples{
equations(
    name = "co", 
    expr = "Qadi+Qbon+Qhea+Qkid+Qliv+Qmus+Qski"
)
}
