% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferencePhysiologies.R
\name{human35m}
\alias{human35m}
\title{Reference physiology of a male adult human}
\usage{
human35m()
}
\description{
Reference physiology of a male adult human
}
\examples{
# Load reference physiology
phys <- human35m()

# Inspect table
phys$param_table

# Export for model
param_list <- to_param_list(phys)
print(param_list)
}
