% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humancovdistrib.R
\name{humancovdistrib}
\alias{humancovdistrib}
\title{Simulate a covariate distribution in adult humans}
\usage{
humancovdistrib(N, sex)
}
\arguments{
\item{N}{Number of individuals (positive integer)}

\item{sex}{Character, either "male" or "female"}
}
\value{
List of \code{Physiology} objects
}
\description{
To generate pairs of (BH,BW) values, independent (BH, BMI) pairs are first
generated, assuming
BH  ~ Normal(muBH,sdBH^2)
BMI ~ logNormal(log_muBMI,log_sdBMI^2)
These are then used to generate BW from the relation BMI = BW/BH^2.
}
\details{
Reference: de la Grandmaison et al., Forensic Sci. Int. 119 (2001): 149-154
}
\examples{
humans <- humancovdistrib(5, "female")
}
