% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartments.R
\name{initials}
\alias{initials}
\title{Initial conditions of a \code{Compartments} object
Returns the initial amounts of compartments in a \code{Compartments} object. The type of the returned object depends
on the presence of units in the initial conditions:}
\usage{
initials(comp, named = FALSE)
}
\arguments{
\item{comp}{A \code{Compartments} object}

\item{named}{If \code{TRUE}, returns a named list of initial amounts; if \code{FALSE} (default), it is unnamed}
}
\value{
A vector of initial amounts for each compartment
}
\description{
\itemize{
\item if all compartment initial conditions are numeric without units, a numeric vector is returned
\item if all compartment initial conditions have consistent units, the returned vector will be of class \code{units},
\item if the compartment initial conditions have mixed units,  the returned vector will be of class \code{mixed_units}.
}
}
