% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{initials}
\alias{initials}
\title{Get initial states of a \code{CompartmentModel} as a named or unnamed vector.}
\usage{
initials(model, named = TRUE)
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}

\item{named}{A boolean, should the initial states be named (default: \code{TRUE})?}
}
\value{
A named or unnamed numeric vector of initial states, in the same order as \code{compartment_names(model)}.
}
\description{
Get initial states of a \code{CompartmentModel} as a named or unnamed vector.
}
