% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartments.R
\name{length.Compartments}
\alias{length.Compartments}
\title{Length method for \code{Compartments} class
Returns the number of compartments in a \code{Compartments} object.}
\usage{
\method{length}{Compartments}(x)
}
\arguments{
\item{x}{A \code{Compartments} object}
}
\value{
The number of compartments (length of the \code{name} vector)
}
\description{
Length method for \code{Compartments} class
Returns the number of compartments in a \code{Compartments} object.
}
