% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dosing.R
\name{length.Dosing}
\alias{length.Dosing}
\title{Length method for Dosing class}
\usage{
\method{length}{Dosing}(x)
}
\arguments{
\item{x}{A \code{Dosing} object}
}
\value{
The number of dosing events (number of rows in the data frame)
}
\description{
Returns the number of dosing events in a \code{Dosing} object.
}
