% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumping.R
\name{lump_model}
\alias{lump_model}
\title{Lump a \code{CompartmentModel} object}
\usage{
lump_model(M, partitioning = list())
}
\arguments{
\item{M}{A \code{CompartmentModel} object, the model to be lumped.}

\item{partitioning}{A list of character arrays, possibly named}
}
\value{
A \code{CompartmentModel} object representing the lumped model
}
\description{
Lump a \code{CompartmentModel} object
}
