% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecules.R
\name{molecules}
\alias{molecules}
\title{\code{Molecules} class
Represents the molecules in a PBPK model.}
\usage{
molecules(molec, cmt = NULL)
}
\arguments{
\item{molec}{Molecule name(s), character scalar or vector}

\item{cmt}{Compartment name(s), character scalar or vector}
}
\value{
A \code{Molecules} object containing the molecule and compartment information for each molecule.
}
\description{
\code{Molecules} class
Represents the molecules in a PBPK model.
}
