% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecules.R
\name{molecules}
\alias{molecules}
\title{\code{Molecules} class
Represents the molecules in a PBPK model.}
\usage{
molecules(
  name = character(0),
  cmt = NULL,
  initial = 0,
  unit = NULL,
  type = c("concentration", "amount")
)
}
\arguments{
\item{name}{Molecule name(s), character scalar or vector}

\item{cmt}{Compartment name(s), character scalar or vector}

\item{initial}{Initial concentration(s)/amount(s) of the molecule (numeric, possibly as \code{values[unit]}). The dimension of the
\code{initial} argument needs to be compatible with the \code{molec} and \code{cmt} arguments (to be clarified).}

\item{unit}{Unit for the initial conditions (character scalar or \code{NULL}, the default), to be used if units are not specified via \code{initial}.}

\item{type}{Type of initial condition, either \code{"concentration"} (the default) or \code{"amount"}.}
}
\value{
A \code{Molecules} object containing the compartment and initial conditioninformation for each molecule.
}
\description{
\code{Molecules} class
Represents the molecules in a PBPK model.
}
\examples{
# No compartment specified: molecules are assumed to be present in all compartments
molecules(name = c("NFkB", "IkB"), unit = "mol/L")
# Compartment specified: each molecule is associated with a specific compartment
molecules(name = c("NFkB", "IkB"), cmt = c("cytoplasm", "nucleus"), initial = c(100, 50) [mol/L])
}
