% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{multiCompModel}
\alias{multiCompModel}
\title{Create a generic multi-compartment PK model}
\usage{
multiCompModel(ncomp = 1, type = c("micro", "macro"))
}
\arguments{
\item{ncomp}{Number of compartments (>=1)}

\item{type}{Style of parameterization: "micro" (rate constants) or "macro" (volumes + clearances)}
}
\value{
A \code{CompartmentModel} object
}
\description{
Create a generic multi-compartment PK model
}
\examples{
multiCompModel(ncomp = 3, type = "macro")
}
