% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartments.R
\name{names.Compartments}
\alias{names.Compartments}
\title{Names method for \code{Compartments} class}
\usage{
\method{names}{Compartments}(x)
}
\arguments{
\item{x}{A \code{Compartments} object}
}
\value{
A character vector of compartment names
}
\description{
Returns the names of compartments in a \code{Compartments} object.
}
