% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{names.Equations}
\alias{names.Equations}
\title{Names method for \code{Equations} objects}
\usage{
\method{names}{Equations}(x)
}
\arguments{
\item{x}{An \code{Equations} object}
}
\value{
The names of the equations
}
\description{
Names method for \code{Equations} objects
}
