% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observable.R
\name{names.Observables}
\alias{names.Observables}
\title{Names method for \code{Observables} objects}
\usage{
\method{names}{Observables}(x)
}
\arguments{
\item{x}{An \code{Observables} object}
}
\value{
The names of the observables
}
\description{
Names method for \code{Observables} objects
}
