% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{names.Parameters}
\alias{names.Parameters}
\title{Names method for \code{Parameters} objects}
\usage{
\method{names}{Parameters}(x)
}
\arguments{
\item{x}{A \code{Parameters} object}
}
\value{
The names of the parameters
}
\description{
Names method for \code{Parameters} objects
}
