% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observable.R
\name{Observable}
\alias{Observable}
\title{Observable Class}
\description{
Observable Class

Observable Class
}
\details{
Represents a model observable
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name of the observable}

\item{\code{expr}}{Expression (as an R call)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Observable-new}{\code{Observable$new()}}
\item \href{#method-Observable-print}{\code{Observable$print()}}
\item \href{#method-Observable-clone}{\code{Observable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Observable-new"></a>}}
\if{latex}{\out{\hypertarget{method-Observable-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new \code{Observable} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observable$new(name, expr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the observable}

\item{\code{expr}}{Expression (character or function)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Observable} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Observable-print"></a>}}
\if{latex}{\out{\hypertarget{method-Observable-print}{}}}
\subsection{Method \code{print()}}{
Print the observable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observable$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments (not used)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{Observable} object (invisible)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Observable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Observable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
