% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observable.R
\name{observables}
\alias{observables}
\title{Create an \code{Observables} object}
\usage{
observables(name, expr)
}
\arguments{
\item{name}{Name of the observable(s), character scalar or vector}

\item{expr}{Expression (as an R call)}
}
\value{
An \code{Observables} object
}
\description{
Create an \code{Observables} object
}
\examples{
observables(
    name = c("Cpla","Cblo"), 
    expr = c("Central/Vcentral/BP", "Central/Vcentral")
)
}
