% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observables.R
\name{observables}
\alias{observables}
\title{Create an \code{Observables} object}
\usage{
observables(..., name = character(0), expr = character(0))
}
\arguments{
\item{...}{Name-expression pairs defining the observables. The expressions can be provided as character strings or as R calls.}

\item{name}{Name of the observable(s), character scalar or vector}

\item{expr}{Expression(s) for the observable(s), as character scalar/vector or as R call / lists of R calls.
Expressions may contain parameters that are added to the model's parameters list, as well as compartment states (not compartment names!)}
}
\value{
An \code{Observables} object
}
\description{
Observables are quantities of interest that are calculated from the state variables
(compartments) and parameters, but are not part of the ODE system.
}
\details{
Observables and equations are very similar, but there are two key differences:
\enumerate{
\item For each observable (but not for equations), \code{to_ode()} returns a function to calculate it based on the ODE output.
\item Equations can be used in the definitions of flows, while observables cannot.
}
}
\examples{
## Interactive path with name-expression pairs
observables(
    Cpla = Acen/Vcen/BP,
    Cblo = Acen/Vcen
)
## Programmatic path with name and expression vectors
observables(
    name = c("Cpla","Cblo"), 
    expr = c("Acen/Vcen/BP", "Acen/Vcen")
)
}
