% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{parameters}
\alias{parameters}
\title{\code{Parameters} class.}
\usage{
parameters(..., name = NULL, value = NULL, unit = NULL)
}
\arguments{
\item{...}{Parameter values, either as named arguments or as a list of named parameters.}

\item{name}{Optional parameter names (if not using named arguments).}

\item{value}{Optional parameter values (if not using named arguments).}

\item{unit}{Optional parameter units (if not using named arguments).}
}
\value{
A \code{Parameters} object containing the specified parameters.
}
\description{
Parameters can be specified in two ways:
\enumerate{
\item As name-value pairs using non-standard evaluation (where each value can optionally have an associated unit).
\item Via the \code{name}, \code{value}, and \code{unit} arguments (where \code{name} is a character vector of parameter names,
\code{value} is a numeric vector of parameter values, and \code{unit} is an optional character vector of units).
}
}
