% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartments.R
\name{print.Compartments}
\alias{print.Compartments}
\title{Print method for \code{Compartments} class
Pretty-prints a \code{Compartments} object.}
\usage{
\method{print}{Compartments}(x, ...)
}
\arguments{
\item{x}{A \code{Compartments} object.}

\item{...}{ignored}
}
\value{
The \code{Compartments} object (invisibly).
}
\description{
Print method for \code{Compartments} class
Pretty-prints a \code{Compartments} object.
}
