% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{print.Equations}
\alias{print.Equations}
\title{Print method for \code{Equations} objects}
\usage{
\method{print}{Equations}(x, ...)
}
\arguments{
\item{x}{An \code{Equations} object}

\item{...}{Additional arguments (not used)}
}
\value{
The \code{Equations} object (invisible)
}
\description{
Print method for \code{Equations} objects
}
