% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessOdeModel.R
\name{print.OdeModel}
\alias{print.OdeModel}
\title{Print method for \code{OdeModel} class}
\usage{
\method{print}{OdeModel}(x, ...)
}
\arguments{
\item{x}{An \code{OdeModel} object.}

\item{...}{ignored}
}
\value{
The \code{OdeModel} object (invisibly).
}
\description{
Pretty-prints an \code{OdeModel} object using DSL state names.
}
