% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{print.Parameters}
\alias{print.Parameters}
\title{Print method for \code{Parameters} objects}
\usage{
\method{print}{Parameters}(x, ...)
}
\arguments{
\item{x}{An \code{Parameters} object}

\item{...}{Additional arguments (not used)}
}
\value{
The \code{Parameters} object (invisible)
}
\description{
Print method for \code{Parameters} objects
}
