% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactions.R
\name{reactions}
\alias{reactions}
\title{\code{Reactions} class and methods}
\usage{
reactions(substrates, products, rate = NULL, const = NULL, cmt = NULL)
}
\arguments{
\item{substrates}{Character vector of substrate molecule names}

\item{products}{Character vector of product molecule names}

\item{rate}{Character string representing the reaction rate (for nonlinear reactions)}

\item{const}{Character string representing the reaction constant (for linear reactions)}

\item{cmt}{Character vector of compartment names where the reaction occurs (optional)}
}
\value{
A \code{Reactions} object containing the reaction information for each reaction.
}
\description{
\code{Reactions} class and methods
}
