% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{residuals.EstimationResult}
\alias{residuals.EstimationResult}
\title{Extract estimation residuals}
\usage{
\method{residuals}{EstimationResult}(object, ...)
}
\arguments{
\item{object}{An \code{EstimationResult}.}

\item{...}{Unused.}
}
\value{
A data frame of observed-minus-fitted residuals.
}
\description{
Extract estimation residuals
}
