% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{sMD_PBPK_12CMT_permbased}
\alias{sMD_PBPK_12CMT_permbased}
\title{12-CMT PBPK model with permeation-based tissue distribution}
\usage{
sMD_PBPK_12CMT_permbased()
}
\value{
An object of class \code{CompartmentModel}
}
\description{
This model implements a 12-compartment PBPK model with permeation-based tissue distribution.
}
\examples{
\donttest{
## Full workflow from model specification to plotting
# Model specification
M <- sMD_PBPK_12CMT_permbased() |>
    add_dosing(time = 0, amount = 1, cmt = "ven")  # IV bolus dosing

paramValues <- list(
    BP = 1,
    CL = 5,
    P = 1,
    fuP = 1,
    fuCadi = 1,
    fuCbon = 1,
    fuCgut = 1,
    fuChea = 1,
    fuCkid = 1,
    fuCliv = 1,
    fuClun = 1,
    fuCmus = 1,
    fuCski = 1,
    fuCspl = 1,
    fuEadi = 1,
    fuEbon = 1,
    fuEgut = 1,
    fuEhea = 1,
    fuEkid = 1,
    fuEliv = 1,
    fuElun = 1,
    fuEmus = 1,
    fuEski = 1,
    fuEspl = 1,
    Qadi = 0.5,
    Qbon = 0.5,
    Qgut = 0.5,
    Qhea = 0.5,
    Qkid = 0.5,
    Qliv = 1,
    Qmus = 0.5,
    Qski = 0.5,
    Qspl = 0.5,
    SAadi = 1,
    SAbon = 1,
    SAgut = 1,
    SAhea = 1,
    SAmus = 1,
    SAkid = 1,
    SAliv = 1,
    SAlun = 1,
    SAski = 1,
    SAspl = 1,
    Vadi_exc = 1,
    Vadi_cel = 1,
    Vbon_exc = 1,
    Vbon_cel = 1,
    Vgut_exc = 1,
    Vgut_cel = 1,
    Vhea_exc = 1,
    Vhea_cel = 1,
    Vmus_exc = 1,
    Vmus_cel = 1,
    Vkid_exc = 1,
    Vkid_cel = 1,
    Vliv_exc = 1,
    Vliv_cel = 1,
    Vlun_exc = 1,
    Vlun_cel = 1,
    Vmus_exc = 1,
    Vmus_cel = 1,
    Vski_exc = 1,
    Vski_cel = 1,
    Vspl_exc = 1,
    Vspl_cel = 1,
    Vart = 1,
    Vven = 1
)

# Run simulation
times <- seq(0, 24, by = 0.1)
out <- simulate(M, time = times, parameters = paramValues)
out_df <- out$states

# Plot compartment amounts
matplot(
    out_df$time,
    out_df[, -1],
    type = "l",
    lty = 1,
    col = 1:12,
    xlab = "Time (h)",
    ylab = "Amount",
    main = "PBPK Model Compartments"
)
legend(
    "topright",
    legend = colnames(out_df)[-1],
    col = 1:12,
    lty = 1
)

obs_df <- out$observables

# Plot plasma concentration
plot(
    obs_df$time,
    obs_df$Cpla,
    type = "l",
    col = "black",
    lwd = 2,
    xlab = "Time (h)",
    ylab = "Plasma concentration",
    main = "Plasma Concentration"
)
}
}
