% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{sMD_PBPK_12CMT_permbased}
\alias{sMD_PBPK_12CMT_permbased}
\title{12-CMT PBPK model with permeation-based tissue distribution}
\usage{
sMD_PBPK_12CMT_permbased()
}
\value{
An R6 object of class \code{CompartmentModel}
}
\description{
This example demonstrates a full workflow from loading the model,
setting parameters, running a simulation, and plotting compartment amounts
and plasma concentration.
}
\examples{
\donttest{
library(compphysiol)
library(deSolve)

# Load the PBPK model
M <- sMD_PBPK_12CMT_permbased()$
    addDosing(target = "ven", time = 0, amount = 1)$  # IV bolus dosing 
    addObservable("Cpla", "BP * ven / Vven")          # plasma concentration observable

paramValues <- list(
    BP = 1,
    CL = 5,
    P = 1,
    fuP = 1,
    fuCadi = 1,
    fuCbon = 1,
    fuCgut = 1,
    fuChea = 1,
    fuCkid = 1,
    fuCliv = 1,
    fuClun = 1,
    fuCmus = 1,
    fuCski = 1,
    fuCspl = 1,
    fuEadi = 1,
    fuEbon = 1,
    fuEgut = 1,
    fuEhea = 1,
    fuEkid = 1,
    fuEliv = 1,
    fuElun = 1,
    fuEmus = 1,
    fuEski = 1,
    fuEspl = 1,
    Qadi = 0.5,
    Qbon = 0.5,
    Qgut = 0.5,
    Qhea = 0.5,
    Qkid = 0.5,
    Qliv = 1,
    Qmus = 0.5,
    Qski = 0.5,
    Qspl = 0.5,
    SAadi = 1,
    SAbon = 1,
    SAgut = 1,
    SAhea = 1,
    SAmus = 1,
    SAkid = 1,
    SAliv = 1,
    SAlun = 1,
    SAski = 1,
    SAspl = 1,
    Vadi_exc = 1,
    Vadi_cel = 1,
    Vbon_exc = 1,
    Vbon_cel = 1,
    Vgut_exc = 1,
    Vgut_cel = 1,
    Vhea_exc = 1,
    Vhea_cel = 1,
    Vmus_exc = 1,
    Vmus_cel = 1,
    Vkid_exc = 1,
    Vkid_cel = 1,
    Vliv_exc = 1,
    Vliv_cel = 1,
    Vlun_exc = 1,
    Vlun_cel = 1,
    Vmus_exc = 1,
    Vmus_cel = 1,
    Vski_exc = 1,
    Vski_cel = 1,
    Vspl_exc = 1,
    Vspl_cel = 1,
    Vart = 1,
    Vven = 1
)

# Generate ODE function and auxiliary structures
odeinfo <- M$toODE(paramValues)

# Run simulation
times <- seq(0, 24, by = 0.1)
out <- ode(
    y = odeinfo$y0,
    times = times,
    func = odeinfo$odefun,
    parms = paramValues,
    events = odeinfo$events
)

out_df <- as.data.frame(out)

# Plot compartment amounts
matplot(
    out_df$time,
    out_df[, -1],
    type = "l",
    lty = 1,
    col = 1:12,
    xlab = "Time (h)",
    ylab = "Amount",
    main = "PBPK Model Compartments"
)
legend(
    "topright",
    legend = colnames(out_df)[-1],
    col = 1:12,
    lty = 1
)

# Evaluate observables at each time point
obs_df <- as.data.frame(sapply(
    odeinfo$obsFuncs,
    function(f) f(out_df$time, as.matrix(out_df[, -1]), paramValues)
))
names(obs_df) <- names(odeinfo$obsFuncs)
obs_df$time <- out_df$time

# Plot plasma concentration
plot(
    obs_df$time,
    obs_df$Cpla,
    type = "l",
    col = "black",
    lwd = 2,
    xlab = "Time (h)",
    ylab = "Plasma concentration",
    main = "Plasma Concentration"
)
}
}
