% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{sMD_PBPK_12CMT_wellstirred}
\alias{sMD_PBPK_12CMT_wellstirred}
\title{12-CMT PBPK model with well-stirred tissue distribution}
\usage{
sMD_PBPK_12CMT_wellstirred()
}
\value{
An object of class \code{CompartmentModel}
}
\description{
This model implements a 12-compartment PBPK model with well-stirred tissue distribution.
}
\examples{
\donttest{
## Full workflow from model specification to plotting
# Model specification
M <- sMD_PBPK_12CMT_wellstirred() |>
    add_dosing(time = 0, amount = 1, cmt = "ven")  # IV bolus dosing

paramValues <- list(
    BP = 1,
    CL = 5,
    Kadi = 1,
    Kbon = 1,
    Kgut = 1,
    Khea = 1,
    Kkid = 1,
    Kliv = 1,
    Klun = 1,
    Kmus = 1,
    Kski = 1,
    Kspl = 1,
    Qadi = 0.5,
    Qbon = 0.5,
    Qgut = 0.5,
    Qhea = 0.5,
    Qkid = 0.5,
    Qliv = 1,
    Qmus = 0.5,
    Qski = 0.5,
    Qspl = 0.5,
    Vadi = 1,
    Vart = 1,
    Vbon = 1,
    Vgut = 1,
    Vhea = 1,
    Vkid = 1,
    Vliv = 1,
    Vlun = 1,
    Vmus = 1,
    Vski = 1,
    Vspl = 1,
    Vven = 1
)

# Run simulation
times <- seq(0, 24, by = 0.1)
out <- simulate(M, time = times, parameters = paramValues)
out_df <- out$states

# Plot compartment amounts
matplot(
    out_df$time,
    out_df[, -1],
    type = "l",
    lty = 1,
    col = 1:12,
    xlab = "Time (h)",
    ylab = "Amount",
    main = "PBPK Model Compartments"
)
legend(
    "topright",
    legend = colnames(out_df)[-1],
    col = 1:12,
    lty = 1
)

obs_df <- out$observables

# Plot plasma concentration
plot(
    obs_df$time,
    obs_df$Cpla,
    type = "l",
    col = "black",
    lwd = 2,
    xlab = "Time (h)",
    ylab = "Plasma concentration",
    main = "Plasma Concentration"
)
}
}
