% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartments.R
\name{[.Compartments}
\alias{[.Compartments}
\title{Subset a \code{Compartments} object}
\usage{
\method{[}{Compartments}(x, i, ...)
}
\arguments{
\item{x}{A \code{Compartments} object}

\item{i}{Index or name(s) of compartments to subset}

\item{...}{Additional arguments (not used)}
}
\value{
A subsetted \code{Compartments} object
}
\description{
Subset a \code{Compartments} object
}
