% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dosing.R
\name{[.Dosing}
\alias{[.Dosing}
\title{Subset method for \code{Dosing} class
Allows subsetting a \code{Dosing} object while preserving its class.}
\usage{
\method{[}{Dosing}(x, i, ...)
}
\arguments{
\item{x}{A \code{Dosing} object}

\item{i}{Row indices to subset}

\item{...}{ignored}
}
\value{
A subsetted \code{Dosing} object
}
\description{
Subset method for \code{Dosing} class
Allows subsetting a \code{Dosing} object while preserving its class.
}
