% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{[.Equations}
\alias{[.Equations}
\title{Subsetting method for \code{Equations} objects}
\usage{
\method{[}{Equations}(x, i, ...)
}
\arguments{
\item{x}{An \code{Equations} object}

\item{i}{Indices or names of the equations to subset}

\item{...}{Additional arguments (not used)}
}
\value{
A subset of the \code{Equations} object
}
\description{
Subsetting method for \code{Equations} objects
}
