% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observables.R
\name{[.Observables}
\alias{[.Observables}
\title{Subsetting method for \code{Observables} objects}
\usage{
\method{[}{Observables}(x, i, ...)
}
\arguments{
\item{x}{An \code{Observables} object}

\item{i}{Indices or names of the observables to subset}

\item{...}{Additional arguments (not used)}
}
\value{
A subset of the \code{Observables} object
}
\description{
Subsetting method for \code{Observables} objects
}
