% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{[.Parameters}
\alias{[.Parameters}
\title{Subsetting method for \code{Parameters} objects}
\usage{
\method{[}{Parameters}(x, i, ...)
}
\arguments{
\item{x}{A \code{Parameters} object}

\item{i}{Indices or names of the parameters to subset}

\item{...}{Additional arguments (not used)}
}
\value{
A subset of the \code{Parameters} object
}
\description{
Subsetting method for \code{Parameters} objects
}
