% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactions.R
\name{[.Reactions}
\alias{[.Reactions}
\title{Subset method for \code{Reactions} class}
\usage{
\method{[}{Reactions}(x, i, ...)
}
\arguments{
\item{x}{A \code{Reactions} object}

\item{i}{Row indices to subset}

\item{...}{Additional arguments (not used)}
}
\value{
A subsetted \code{Reactions} object
}
\description{
Allows subsetting a \code{Reactions} object while preserving its class.
}
