% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumping.R
\name{symbolic_lumping}
\alias{symbolic_lumping}
\title{General symbolic lumping (experimental version)}
\usage{
symbolic_lumping(M, partitioning = list(), ref = .get_refstate(M))
}
\arguments{
\item{M}{A CompartmentModel object}

\item{partitioning}{A list of character vectors (each a group of compartments)}

\item{ref}{A string specifying a reference compartment (defaults to the one with most outflows)}
}
\value{
A CompartmentModel object representing the lumped model
}
\description{
General symbolic lumping (experimental version)
}
