% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{to_analytical}
\alias{to_analytical}
\title{Generate analytical solution function from a linear model object.}
\usage{
to_analytical(model)
}
\arguments{
\item{model}{A \code{CompartmentModel} or \code{AnalyticalModel} object.}
}
\value{
A length 2 list named \code{state} (a function) and \code{observable}
(a list of functions, possibly empty). \code{state(t,param)} calculates the ODE
solution at time \code{t} for free parameters \code{param}, while \code{observable[[i]](t,param)}
calculates the \code{i}-th observable defined in the CompartmentModel.
}
\description{
For linear models, the system of ODEs can be solved analytically using
matrix exponentials. This function generates a state function that evaluates
this analytical solution at given time points and parameter values.
}
\details{
Analytical export first lowers \code{CompartmentModel} inputs to an
\code{AnalyticalModel}. Passing an \code{AnalyticalModel} directly avoids repeated
lowering when exporting the same model structure repeatedly.
}
\examples{
M <- multiCompModel(ncomp = 2, type = "micro") |>
    add_parameter(kc0 = 0.05)
sol <- to_analytical(M)

# Evaluate ODE state at t = 5 with free params
sol$statefun(5, params = list(kcp = 0.2, kpc = 0.1))
}
