% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{to_analytical}
\alias{to_analytical}
\title{Generate analytical solution function from a linear \code{CompartmentModel} object, possibly with a single bolus dose at time 0.}
\usage{
to_analytical(model, paramValues = list())
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}

\item{paramValues}{Named list of parameter values to inline in ODEs.}
}
\value{
A length 2 list named \code{state} (a function) and \code{observable}
(a list of functions, possibly empty). \code{state(t,param)} calculates the ODE
solution at time \code{t} for free parameters \code{param}, while \code{observable[[i]](t,param)}
calculates the \code{i}-th observable defined in the CompartmentModel.
}
\description{
For linear compartment models, the system of ODEs can be solved analytically using matrix exponentials.
This function generates a state function that evaluates this analytical solution at given time points and parameter values.
}
\examples{
M <- multiCompModel(ncomp = 2, type = "micro")
sol <- to_analytical(M, paramValues = list(k10 = 0.05))  # fix one param

# Evaluate ODE state at t = 5 with free params
sol$statefun(5, params = list(k12 = 0.2, k21 = 0.1))
}
