% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessOdeModel.R
\name{to_deSolve}
\alias{to_deSolve}
\title{Export an ODE model to deSolve format}
\usage{
to_deSolve(model, parameters = list(), dimensions = NULL)
}
\arguments{
\item{model}{An \code{OdeModel} object.}

\item{parameters}{Runtime parameter overrides.}

\item{dimensions}{Named list of unit dimensions.}
}
\value{
A list suitable for \code{\link[deSolve:ode]{deSolve::ode()}}.
}
\description{
Export an ODE model to deSolve format
}
