% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompartmentModel.R
\name{to_ode}
\alias{to_ode}
\title{Generate ODE function, initial values, observables, and free parameters from a \code{CompartmentModel} object.}
\usage{
to_ode(model, paramValues = list())
}
\arguments{
\item{model}{A \code{CompartmentModel} object.}

\item{paramValues}{Named list of parameter values to inline in ODEs.}
}
\value{
A list with elements \code{odefun} (function), \code{y0} (named numeric vector), \code{obsFuncs} (list of functions),
and \code{freeParams} (character vector).
}
\description{
Generate ODE function, initial values, observables, and free parameters from a \code{CompartmentModel} object.
}
\examples{
M <- multiCompModel(ncomp = 2, type = "micro")
odeinfo <- to_ode(M, paramValues = list(k10 = 0.05))
}
