% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{to_param_list}
\alias{to_param_list}
\title{Export parameters as a named list
Scalars: name → value
Tissues: \code{paste0(parameter, "[", context, "]")} → value}
\usage{
to_param_list(phys)
}
\arguments{
\item{phys}{A \code{Physiology} object}
}
\value{
A named list of parameters and metadata
}
\description{
Export parameters as a named list
Scalars: name → value
Tissues: \code{paste0(parameter, "[", context, "]")} → value
}
