% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transports.R
\name{transports}
\alias{transports}
\title{Transports Class}
\usage{
transports(from, to, molec = NA_character_, ..., rate = NULL, const = NULL)
}
\arguments{
\item{from}{The name of the source compartment(s) (character vector or NULL for source compartments)}

\item{to}{The name of the destination compartment(s) (character vector or NULL for sink compartments)}

\item{molec}{The name of the molecule(s) being transferred (character vector, optional, default: all molecules)}

\item{...}{Unused, enforces \code{rate} and \code{const} to be specified as named arguments only, not positional}

\item{rate}{The transport rate (character or call, optional, mutually exclusive with \code{const})}

\item{const}{Rate constant for first-order transports (character or call, optional, mutually exclusive with \code{rate})}
}
\value{
A \code{Transports} object
}
\description{
Transports represent mass/amount-preserving transports of molecules between two compartments in a compartment model.
}
\details{
Transports can be linear (first-order) or nonlinear, depending on whether the transport rate is defined by a constant
(argument \code{const}) or a general expression (argument \code{rate}).
The special syntax \code{{from}}, \code{{to}} and \code{{molec}} can be used in \code{rate} or \code{const} to automatically substitute
the source and destination compartment names and molecule names, respectively, e.g., \code{"k_{from}_{to}"} will become
\code{k_ven_bon} for a transport from the venous blood (ven) to bone (bon). This allows for vector definition of multiple
transports with different compartments and/or molecules involved.
}
\examples{
# Linear transport
f1 <- transports(from = "A", to = "B", const = "k1")   # rate will be k1 * a[A]
# Nonlinear transport
f2 <- transports(from = "A", to = "B", rate = "k1 * A*B/(B+K)")
}
