% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{unlist.Parameters}
\alias{unlist.Parameters}
\title{Unlist a \code{Parameters} object (experimental).}
\usage{
\method{unlist}{Parameters}(x, ...)
}
\arguments{
\item{x}{A \code{Parameters} object}

\item{...}{Additional arguments (not used)}
}
\value{
An unlisted \code{Parameters} object
}
\description{
Parameters may be scalars or vectors, but for some purposes it may be desirable to unlist them into a flat \code{Parameters} object.
}
